% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addRuleQuality}
\alias{addRuleQuality}
\title{Add quality measures to association rules}
\usage{
addRuleQuality(trans, rules, include = NULL, exclude = "improvement")
}
\arguments{
\item{trans}{transaction set (s4 class from arules package)}

\item{rules}{set of rules from \code{apriori}}

\item{include}{character vector specifying which quality measures to include.  Default is to include everything.}

\item{exclude}{character vector specifying which quality measures to exclude. Default is to exclude 'improvement' because it seems to be slow.}
}
\value{
ruleset with additional quality measures
}
\description{
Adds measures of rule quality (conviction, hyperConfidence, cosine, chiSquare, coverage, doc, gini, hyperlift) to a set of
association rules mined from \code{apriori}.  Usually used before converting ruleset to data.frame and exporting to some sort of text file.
}
\examples{
library('arules')
data("Adult")
ar <- apriori(Adult, parameter = list(supp = 0.5, conf = 0.9, target = "rules"))
ar <- addRuleQuality(trans=Adult, rules=ar)
df <- Rsenal::rules2df(ar)
}

