% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/borg_indices.R
\name{borg_indices}
\alias{borg_indices}
\title{Calculate shape indices listed in Borg 1998}
\usage{
borg_indices(Ar, Ur, De, Du)
}
\arguments{
\item{Ar}{area of the entities}

\item{De}{Diameter of the inner circle}

\item{Du}{Diameter of the outer circle}
}
\value{
A RasterStack
}
\description{
used as internal funcction. see \code{?geometry_variables}
and Borg 1998: Vergleichende Analyse von Formindizes zur Characterisierung
von Landschaftsobjekten unter ökologischen Aspekten.
Zeitschrift für Photogrammetrie und Fernerkundung 4: 108-119.
}
\author{
Hanna Meyer
}
\seealso{
\code{?geometry_variables}
}

