% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borgIndices.R
\name{borgIndices}
\alias{borgIndices}
\title{Calculate shape indices listed in Borg 1998}
\usage{
borgIndices(Ar, Ur, De, Du)
}
\arguments{
\item{Ar}{area of the entities}

\item{De}{Diameter of the inner circle}

\item{Du}{Diameter of the outer circle}
}
\value{
A RasterStack
}
\description{
used as internal funcction. see \code{?geometry_variables}
and Borg 1998: Vergleichende Analyse von Formindizes zur Characterisierung 
von Landschaftsobjekten unter ökologischen Aspekten. 
Zeitschrift für Photogrammetrie und Fernerkundung 4: 108-119.
}
\author{
Hanna Meyer
}
\seealso{
\code{?geometryVariables}
}

