% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bss.R
\name{bss}
\alias{bss}
\title{Best subset feature selection}
\usage{
bss(predictors, response, method = "rf",
  metric = ifelse(is.factor(response), "Accuracy", "RMSE"),
  maximize = ifelse(metric == "RMSE", FALSE, TRUE),
  trControl = trainControl(), tuneLength = 3, tuneGrid = NULL,
  seed = 100)
}
\arguments{
\item{predictors}{see \code{\link{train}}}

\item{response}{see \code{\link{train}}}

\item{method}{see \code{\link{train}}}

\item{metric}{see \code{\link{train}}}

\item{maximize}{see \code{\link{train}}}

\item{trControl}{see \code{\link{train}}}

\item{tuneLength}{see \code{\link{train}}}

\item{tuneGrid}{see \code{\link{train}}}

\item{seed}{A random number}
}
\value{
A list of class train
}
\description{
Evaluate all combinations of predictors during model training
}
\details{
Models are iteratively fitted using all different combinations
of predictor variables. Thus 2^X models are calculated.

The internal cross validation can be run in parallel. See information
on parallel processing of carets train functions for details.
}
\note{
This validation is particulary suitable for 
leave-one-station-out cross validations where variable selection
MUST be based on the performance of the model on the hold out station.
A more time efficient alternative is the forward feature selection
 (\code{\link{ffs}}).
}
\examples{
#' \dontrun{
data(iris)
bssmodel <- bss(iris[,1:4],iris$Species)
bssmodel$finalModel$xNames
}
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{train}},  \code{\link{ffs}}, 
\code{\link{trainControl}},\code{\link{rfe}}
}

