% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bumpVersion}
\alias{bumpVersion}
\title{bump package 'Version:' and 'Date:' in DESCRIPTION file}
\usage{
bumpVersion(major, minor, patch, pkg.repo = ".")
}
\arguments{
\item{major}{major versioning number (numeric)}

\item{minor}{minor versioning number (numeric)}

\item{patch}{patch versioning number (numeric)}

\item{pkg.repo}{path to package repository folder. Default is current
working directory}
}
\description{
this function let's you bump the version number and creation date of
your package's DESCRIPTION file.
Supported versioning system is major.minor.patch
}
\author{
Tim Appelhans
}

