% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcOffsetGridText.R
\name{calcOffsetGridText}
\alias{calcOffsetGridText}
\title{Calculate offset coordinates for grid-based text annotations}
\usage{
calcOffsetGridText(x, y = NULL, offset = 0.02, pos = NULL, xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{x}{Numeric. A numeric vector containing x coordinates, or a 2-column
numeric matrix containing x and y coordinates.}

\item{y}{Numeric. A numeric vector containing y coordinates, or \code{NULL}
if x is a two-column matrix.}

\item{offset}{Numeric. The desired offset in normalized parent coordinates
("npc", see \code{\link{unit}}).}

\item{pos}{Integer. Text position specifier(s) as used by \code{\link{text}}.
If not supplied, optimal text positions will be determined with respect to
neighboring locations using \code{\link{thigmophobe}}.}

\item{xlim}{Numeric. X-axis limits (xmin, xmax) of the current plot. If not
supplied, limits are automatically calculated from supplied x and y
coordinates.}

\item{ylim}{Numeric. Y-axis limits (ymin, ymax) of the current plot. If not
supplied, limits are automatically calculated from supplied x and y
coordinates.}

\item{...}{Further arguments. Currently not in use.}
}
\value{
A numeric matrix containing offset coordinates.
}
\description{
Calculate offset coordinates for (\strong{grid}-based) text drawing
functions, e.g. \code{\link{grid.text}}.
}
\examples{
# required packages
library(grid)
library(plotrix)

# sample point data
x_cntr <- 330000
y_cntr <- 9650000

ls_xy <- lapply(c(x_cntr, y_cntr), function(i) {
  set.seed(i * 100)
  rnorm(10, i, 10000)
})

df <- data.frame(id = 1:10,
                 x = ls_xy[[1]],
                 y = ls_xy[[2]])
coordinates(df) <- ~ x + y
projection(df) <- "+init=epsg:21037"

# offset point locations
loc_lbl <- calcOffsetGridText(x = coordinates(df), offset = .01)

# vis
spplot(df, col.regions = "black", auto.key = FALSE)

downViewport(trellis.vpname(name = "figure"))
for (i in 1:nrow(df))
  grid.text(label = df$id[i], x = loc_lbl[i, 1], y = loc_lbl[i, 2],
            just = pos2just(thigmophobe(coordinates(df)))[i])
}
\author{
Florian Detsch
}
\seealso{
\code{\link{grid.text}}, \code{\link{text}}, \code{\link{thigmophobe}}
}

