% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charCloud.R
\name{charCloud}
\alias{charCloud}
\title{create a word cloud from a character string}
\usage{
charCloud(txt, exclude = NULL, nwords = 180, min.char.length = 5,
  max.char.length = 30, clrs = c("#8c510a", "#bf812d", "#35978f", "#01665e",
  "#7fbc41", "#4d9221"), seed = 123, wmin = NULL, stat = FALSE,
  sortby = c("freq, names"), ...)
}
\arguments{
\item{txt}{character string from which the cloud will be computed}

\item{exclude}{optional path to a .txt file with words to be excluded 
from the cloud}

\item{nwords}{the number of words to be plotted}

\item{min.char.length}{minimum length of words to be included in the cloud}

\item{max.char.length}{maximum length of words to be included in the cloud}

\item{clrs}{vector of colors to be used for the word cloud}

\item{seed}{seed to be used for the random word placement}

\item{wmin}{see \code{\link{tagcloud}} for details}

\item{stat}{return statistics (words and frequencies)}

\item{sortby}{if stat is TRUE, sort statistics by names or frequency}

\item{...}{additional arguments passed to \code{\link{tagcloud}}}
}
\description{
this function creates a wordcloud based on a character string. 
Optionally, a file with common words to be excluded can be 
supplied. See \code{\link{tagcloud}} for details on the cloud building.
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{tagcloud}}
}

