% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr2Geos.R
\name{cr2Geos}
\alias{cr2Geos}
\title{column/row coordinates to geos projection}
\usage{
cr2Geos(cr, COFF = 1856, ColumnDirGridStep = 3000.40316581726)
}
\arguments{
\item{cr}{Numeric value representing the column/row coordinate of a MSG image}
}
\value{
Numeric value representing the column/row coordinate in Geos projection
}
\description{
Converts column/row coordinates of MSG images to geos projection.
For HRV use COFF=5566 and ColumnDirGridStep=1000.13434886932 instead of the 
default values.
}
\examples{
cr2Geos(2325)

}
\author{
Hanna Meyer and Johannes Droenner
}

