% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createHdr}
\alias{createHdr}
\title{Create ENVI header file}
\usage{
createHdr(hdr = NULL, file = "./tmp.hdr", ...)
}
\arguments{
\item{hdr}{Character. If not supplied, defaults to the companion header file
for GIMMS 3G binary data.}

\item{file}{Character. Filepath passed on to \code{\link{writeLines}}.}

\item{...}{Further arguments. Currently not in use.}
}
\value{
A filename corresponding to \code{file} argument.
}
\description{
Create an ENVI header file (see \url{http://www.exelisvis.com/docs/ENVIHeaderFiles.html})
following the documentation for \code{\link{read.ENVI}}.
}
\examples{
# Create standard GIMMS 3G header file in current working directory
createHdr()
}
\author{
Florian Detsch
}
\seealso{
\code{\link{read.ENVI}}
}

