% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createTimeSeries}
\alias{createTimeSeries}
\title{create a regular time series}
\usage{
createTimeSeries(start, end, step, type = c("POSIX", "char"))
}
\arguments{
\item{start}{character string of starting date and time. Needs to be of
format "YYYY-mm-dd HH:MM:SS"}

\item{end}{character string of ending date and time. Needs to be of
format "YYYY-mm-dd HH:MM:SS"}

\item{step}{the time step of the time series (in seconds)}
}
\value{
a character vector of datetimes spaced regularly from start to end
}
\description{
This function will create a (empty) character time series.
This can be used for creating regular time series of real-world observations
that have missing data via merging.
}
\examples{
## 24 hrs
createTimeSeries("1990-01-01 00:00:00", "1990-01-01 23:00:00", step = 60 * 60)

## 365 days
createTimeSeries("1990-01-01 00:00:00", "1990-12-31 23:00:00", step = 60 * 60 * 24)
}

