% Generated by roxygen2 (4.0.1): do not edit by hand
\name{downloadGimms}
\alias{downloadGimms}
\title{Download GIMMS 3G data}
\usage{
downloadGimms(decades, dsn = ".", ...)
}
\arguments{
\item{decades}{Numeric. Decadal data to download. If not supplied, the entire
GIMMS 3G dataset will be acquired, i.e. \code{decades = seq(1980, 2010, 10)}.}

\item{dsn}{Character. Destination folder for file download. If not supplied,
all downloaded files will be stored in the current working directory.}

\item{...}{Further arguments. Currently not in use.}
}
\value{
A vector of filepaths.
}
\description{
Download GIMMS 3G binary data for a given time span from NASA FTP server
(\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g/}).
}
\examples{
# Download GIMMS 3G data from 1990 to 2009 (this might take some time...)
downloadGimms(decades = c(1990, 2000),
              dsn = paste(getwd(), "data", sep = "/"))
}
\author{
Florian Detsch
}
\seealso{
\code{\link{download.file}}
}

