% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTRMM.R
\name{downloadTRMM}
\alias{downloadTRMM}
\title{Download TRMM 3B42 Daily Data}
\usage{
downloadTRMM(begin, end, dsn = ".", xml = FALSE, overwrite = FALSE,
  cores = 1L, ...)
}
\arguments{
\item{begin, end}{Start and end date as \code{Date} or \code{character}.}

\item{dsn}{\code{character}. Download folder, defaults to the current working 
directory.}

\item{xml}{\code{logical}, defaults to \code{FALSE}. If \code{TRUE}, .xml 
files associated with each .nc4 file are also downloaded to 'dsn'.}

\item{overwrite}{\code{logical}, defaults to \code{FALSE}. Determines whether 
existing files are overwritten.}

\item{cores}{\code{integer}. Number of cores for parallel processing. Note 
that this takes only effect if a sufficiently fast internet connection is 
available.}

\item{...}{Additional arguments passed to \code{\link{as.Date}}.}
}
\value{
A \code{character} vector of filepaths.
}
\description{
Download TRMM 3B42 daily NetCDF files for a given time span from the NASA FTP 
servers (\url{ftp://disc3.nascom.nasa.gov/data/s4pa/TRMM_L3/TRMM_3B42_Daily.7/}).
}
\examples{
 
\dontrun{
## download TRMM 3B42 data from Jan 1 to Jan 5, 2015
downloadTRMM(begin = "2015-01-01", end = "2015-01-05")
}
              
}
\author{
Florian Detsch
}
\seealso{
\code{\link{download.file}}.
}

