% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/envinmr.theme.R
\name{envinmr.theme}
\alias{envinmr.theme}
\title{Environmental Informatics Marburg lattice plotting theme}
\usage{
envinmr.theme(win.fontfamily = NULL, with.bg = FALSE, box = "black", ...)
}
\description{
This theme is a modified version of \code{\link{theEconomist.theme}}.
Especially the regions colour setting has been modified to a
conceptually pleasant rainbow colour palette (optimised for temperature).
}
\examples{
library(lattice)
levelplot(matrix(rnorm(10000), 100, 100),
          par.settings = envinmr.theme(), at = seq(-5, 5, 0.1))
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{theEconomist.theme}}
}

