% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/envinmrRasterPlot.R
\name{envinmrRasterPlot}
\alias{envinmrRasterPlot}
\title{modify spplot objects}
\usage{
envinmrRasterPlot(spplot.obj, rot = 90, col, ...)
}
\arguments{
\item{spplot.obj}{the \code{\link{spplot}} object to be modified}

\item{rot}{Rotation angle of the y-axis tick labels. Defaults to 90 degrees.}

\item{col}{Color vector. If not specified, color information will
automatically be extracted from \code{spplot.obj}.}

\item{...}{additional arguments passed to \code{\link{update.trellis}}}
}
\value{
a single lattice plot object
}
\description{
this function updates existing \code{\link{spplot}} objects so that
i) the axes are drawn at all four sites (as these usually represent
coordinates) and ii) the colorkey is plotted on top of the graph so
that the main = ... argument can be used to describe the colorkey.
}
\examples{
rst <- raster::raster(volcano)

p1 <- spplot(rst)
p1

p2 <- envinmrRasterPlot(p1, col.regions = envinmrPalette(100))
p2

## change the colors as you like
p3 <- envinmrRasterPlot(p2,
                        col.regions =
                         colorRampPalette(brewer.pal(9, "YlGnBu"))(100))
p3

latticeCombineGrid(list(p2, p1, p3))
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{update.trellis}}, \code{\link{spplot}}
}

