% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extendByDist.R
\name{extendByDist}
\alias{extendByDist}
\title{Extend an object's extent by distance}
\usage{
extendByDist(x, dist)
}
\arguments{
\item{x}{Extent, Raster* or Spatial* object.}

\item{dist}{Numeric. Distance (in native map units) from original geometry.}
}
\value{
An extended extent.
}
\description{
Similar to \code{\link{extend}}, this function extends the extent of an
Extent, Raster* or Spatial* object. Rather than requiring the user
to supply either a target extent or a desired number of columns and rows to
add to the initial Raster*, extension is based on native map units, e.g.
decimal degrees when the projection is EPSG:4326.
}
\examples{
rst <- kiliAerial(rasterize = TRUE, minNumTiles = 4L)

extent(rst)
extendByDist(rst, dist = 1000)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{extend}}, \code{\link{extent}}
}

