% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffs.R
\name{ffs}
\alias{ffs}
\title{Forward feature selection}
\usage{
ffs(predictors, response, method = "rf",
  metric = ifelse(is.factor(response), "Accuracy", "RMSE"),
  maximize = ifelse(metric == "RMSE", FALSE, TRUE),
  trControl = trainControl(), tuneLength = 3, tuneGrid = NULL,
  seed = 100, runParallel = FALSE)
}
\arguments{
\item{predictors}{see \code{\link{train}}}

\item{response}{see \code{\link{train}}}

\item{method}{see \code{\link{train}}}

\item{metric}{see \code{\link{train}}}

\item{maximize}{see \code{\link{train}}}

\item{trControl}{see \code{\link{train}}}

\item{tuneLength}{see \code{\link{train}}}

\item{tuneGrid}{see \code{\link{train}}}

\item{seed}{A random number}

\item{runParallel}{Logical}
}
\value{
A list of class train
}
\description{
A simple forward feature selection algorithm
}
\details{
Models with two predictors are first trained using all possible 
pairs of predictor variables. The best model of these initial models is kept.
On the basis of this best model the predictor variables are iteratively
increased and each of the remaining variables is tested for its improvement
of the currently best model. The process stops if none of the remaining 
variables increases the model performance when added to the current best model.

The internal cross validation can be run in parallel. See information
on parallel processing of carets train functions for details.
}
\note{
This validation is particulary suitable for 
leave-one-station-out cross validations where variable selection
MUST be based on the performance of the model on the hold out station.
A computational time expesnsive alternative is the best subset
selection (\code{\link{bss}}).
}
\examples{
#' \dontrun{
data(iris)
ffsmodel <- ffs(iris[,1:4],iris$Species)
ffsmodel$finalModel$xNames
}
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{train}},  \code{\link{bss}}, 
\code{\link{trainControl}},\code{\link{rfe}}
}

