% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/figureCaption.R
\name{figureCaption}
\alias{figureCaption}
\title{Add a caption to a lattice or ggplot2 object}
\usage{
figureCaption(obj, txt, ...)
}
\arguments{
\item{obj}{the \link{lattice} or \link{ggplot2} to which the caption
should be added}

\item{txt}{the cpation text}

\item{...}{additional parameters passed on to \code{\link{gpar}}. These
can be used to adjust the fontsize, fontfamily etc. Note that the
default value for fontsize is 16 pt.}
}
\description{
this function adds a caption to an existing plot object
(either \link{lattice} or \link{ggplot2}). \code{\link{strwrap}}
is used to automatically wrap text at column 75. Clipping is off,
so long text will continue downward outside the plotting region.
Keep this in mind when setting the height of your graphics device.
}
\examples{
## lattice example
library(lattice)

p1 <- levelplot(volcano)
txt <- "Fig 1.: Elevation of Maunga Whau (Mt Eden) volcano in Auckland.
Axes are standardised values."

figureCaption(p1, txt)

## change font size
figureCaption(p1, txt, fontsize = 12)

## ggplot2 example
p2 <- qplot(1:10, rnorm(10), colour = runif(10))
figureCaption(p2, "Fig. 2: Some color coded random numbers.",
              fontface = "italic")
}
\author{
Tim Appelhans
}

