% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterRules}
\alias{filterRules}
\title{Post process association rules}
\usage{
filterRules(rules, oneSideOnly = oneSideOnly)
}
\arguments{
\item{rules}{list of association rules (S4 arules object).  Output of \code{apriori} function.}

\item{oneSideOnly}{list of character vectors.  Each character vector of the list contains variables that are likely very similar
and will generate uninteresting rules.  So the filtering algorithm will prune rules where variables within this list appear on
both the RHS & LHS}
}
\value{
association rules
}
\description{
This function filters rules that have already been mined.
}
\examples{
library('arules')
data('Adult')
rules <- apriori(Adult, parameter = list(supp=0.01, conf=0.9, target = "rules"))
oneSideOnly <- list(c('age', 'workclass', 'education'), c('marital-status', 'occupation', 'race'))
f2 <- filterRules(rules, oneSideOnly=oneSideOnly)
length(rules)
length(f2)
}

