% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fixWidth.R
\name{fixWidth}
\alias{fixWidth}
\title{fixed width numeric/integer to character conversion}
\usage{
fixWidth(x, n = 1L)
}
\arguments{
\item{x}{a numeric or integer vector}

\item{n}{the width (nchar) of the resulting character}
}
\value{
A character vector of length x
}
\description{
this function is a wrapper around a \code{\link{sprintf}} call for
numeric and integer values (I seem to be unable to memorise the sprintf
call, hence the wrapper).
}
\examples{
fixWidth(1, 2)
fixWidth(c(1, 20, 300, 4000, 50000), 5)
}
\author{
Tim Appelhans
}

