% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funN.R
\name{funN}
\alias{funN}
\title{create a function conditional to a set number of observations}
\usage{
funN(n, fun, ...)
}
\arguments{
\item{n}{the number of minimum observations for evaluation}

\item{fun}{the function to be evaluated}

\item{...}{additional arguments passed to the function (e.g. na.rm = )}
}
\value{
a function that will calculate \code{fun} on a given vector if at least
\code{n} valid observations are present
}
\description{
This function is a so-called closure. It will set up a function so that
it will only be evaluated if the vector on which it is called has at least
\code{n} observations. If it doesn't, it will return NA.
}
\examples{
## create some dummy data
vec <- rnorm(100, 5, 1)
vec[sample(length(vec), 20)] <- NA #replace 20 values of vec with NA

f <- funN(70, mean, na.rm = TRUE)
f(vec)

f1 <- funN(90, mean, na.rm = TRUE)
f1(vec)

f3 <- funN(80, quantile, na.rm = TRUE)
f3(vec)

f4 <- funN(81, quantile, na.rm = TRUE)
f4(vec)
}
\author{
Tim Appelhans
}

