% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryVariables.R
\name{geometryVariables}
\alias{geometryVariables}
\title{Calculate selected geometry variables from raster data}
\usage{
geometryVariables(x)
}
\arguments{
\item{x}{A rasterLayer with any values but where patches are seperated by surrounding NA values}
}
\value{
A RasterStacks containing the geoemtry parameters. The names are as following;
Patches: The ID of the Patches
Ar: The area of the individual patches
SI: The shape index calculated by SDMTools
CA: The core area calculated by SDMTools
Up: The perimeter of the indivisual patches
CAI: The core area index calculated by SDMTools
PAR:The perimeter area ratio calculated by SDMTools
Re: The radius of the inner circle
distEdges: The distcance from edges
Ru: The radius of the outer circle
OIC: The inner-outer-circle ratio
CI1: inner compactness
CO1: outer compactness
CI2: inner compactness2
CO2: outer compactness2
CCI1: corrected compactness index 1
CCI2: corrected compactness index 2
CO: compactness
SHD: Shoreline development
C1: Circularity
E: Elongateness
TR:Thinness Ratio
CR: Circularity ratio
C2: Circularity2
FR: Form ratio
EI: Elipticity index
SF1: Shape factor
GSI: Grain shape index
SF2: Shape factor 2
C3: Circularity 3
SF3: Shape factor 3

For decription of the indices see list in Borg 1998: Vergleichende Analyse von Formindizes zur Characterisierung
von Landschaftsobjekten unter ökologischen Aspekten.
Zeitschrift für Photogrammetrie und Fernerkundung 4: 108-119
}
\description{
This function calculates geoemtry parameters from Raster data.
}
\examples{
RasterLayer<-raster("inst/training.rst")
shape<-geometryVariables(RasterLayer)
plot(shape)
}
\author{
Hanna Meyer
}
\seealso{
\code{SDMTools},  \code{clump}, \code{borg_indices}
}

