% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getEfromVlist}
\alias{getEfromVlist}
\title{Get edge IDs from vertices}
\usage{
getEfromVlist(g, vlist, mode = "out")
}
\arguments{
\item{g}{igraph object}

\item{vlist}{list of node/vertex names}

\item{mode}{character string: 'in', 'out' or 'all'}
}
\value{
edge IDs that go OUT or IN from the specified list of vertices
}
\description{
returns the edge IDs that go OUT or IN from the specified list of vertices
}
\examples{
require('igraph')

## build sample network
from <- c('A','A','B','B','C','C','D','D','E','E','E','F','F','H','L','L','O')
to <- c('B','C','D','E','F','G','H','I','J','J','K','L','G','M','N','O','P')
relations<-cbind(from,to)
g <- graph.data.frame(relations)
tiers <- c(1,2,3,4,1,3,4,2,1,2,3,4,0,0,1,2)
V(g)$tier <- tiers

getEfromVlist(g, c('A', 'D'), 'to')
}

