% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getMode}
\alias{getMode}
\title{compute the mode of column within a data frame}
\usage{
getMode(df, col, side = c("lower", "upper"))
}
\arguments{
\item{df}{a data frame}

\item{col}{a column name or column number}

\item{side}{(requires documentation)}
}
\value{
a vector
}
\description{
this function computes the modal value of a specified column of a data frame
}
\examples{
df <- data.frame(a = c(1, 2, 3, 3, 3, 3, 4, 5, 6, 7), 
                 b = c("a", "b", "c", "d", "e", "e", "e", "e", "f", "g"))

getMode(df, "a")
getMode(df, 2)

}
\author{
Tim Appelhans, Florian Detsch
}

