% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTRMMDateTime.R
\name{getTRMMDateTime}
\alias{getTRMMDateTime}
\alias{getTRMMDateTime}
\alias{getTRMMDate}
\alias{getTRMMTime}
\title{TRMM 3B42 Date/Time Functions}
\usage{
getTRMMDateTime(x, as_posix = FALSE)

getTRMMDate(x, as_date = FALSE)

getTRMMTime(x)
}
\arguments{
\item{x}{\code{character}. TRMM 3B42 filename(s).}

\item{as_posix, as_date}{\code{logical}, default to \code{FALSE}. See 
respective function descriptions below.}
}
\value{
See 'Functions' section below.
}
\description{
Extract date/time information from TRMM 3B42 files.
}
\section{Functions}{
\itemize{
\item \code{getTRMMDateTime}: Get datetime of 3-hourly TRMM 3B42 file(s) 
(\code{character} or \code{POSIXt} depending on 'as_posix').

\item \code{getTRMMDate}: Get date of TRMM 3B42 file(s) (\code{character} 
or \code{Date} depending on 'as_date').

\item \code{getTRMMTime}: Get time of 3-hourly TRMM 3B42 file(s) 
(\code{character}).
}}

\examples{
getTRMMDateTime("3B42.20150106.03.7.HDF")
getTRMMDateTime("3B42.20150106.03.7.HDF", as_posix = TRUE)
    
            
}
