% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gisView.R
\docType{methods}
\name{gisView}
\alias{gisView}
\alias{gisView,RasterStack-method}
\title{view raster layers interactively}
\usage{
\S4method{gisView}{RasterLayer}(x, map = NULL, cols = envinmrPalette(7),
  na.color = "transparent", values = NULL, map.types = c("OpenStreetMap",
  "Esri.WorldImagery"), layer.opacity = 0.8, legend = TRUE,
  legend.opacity = 1, trim = TRUE, ...)

\S4method{gisView}{RasterStack}(x, ...)
}
\arguments{
\item{x}{a \code{\link{raster}}* object}

\item{cols}{color palette for the layers}

\item{na.color}{color for missing values}

\item{values}{a vector of values for the visualisation of the layers.
Per default these are calculated based on the supplied raster* object.}

\item{map.types}{character spcifications for the base maps.
see \url{http://leaflet-extras.github.io/leaflet-providers/preview/}
for available options.}

\item{layer.opacity}{opacity of the layers}

\item{legend}{should a legend be plotted}

\item{legend.opacity}{opacity of the legend}

\item{trim}{should the raster be trimmed in case there are NAs on the egdes}

\item{...}{additional arguments passed on to \code{\link{addLegend}}}
}
\description{
this function produces an interactive GIS-like view of the specified
raster layers on top of the specified base maps.
}
\section{Methods (by class)}{
\itemize{
\item \code{RasterStack}: 
}}
\examples{
LUC1990 <- raster(system.file("LUC1990.rst", package = "Rsenal"))
LUC2006 <- raster(system.file("LUC2006.rst", package = "Rsenal"))

LUC1990_f <- as.factor(LUC1990)
LUC2006_f <- as.factor(LUC2006)

gisView(LUC1990_f)
gisView(LUC2006_f)

stck <- stack(LUC1990_f, LUC2006_f)
gisView(stck)
}
\author{
Tim Appelhans
}

