% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsenal-package.R
\docType{data}
\name{gmap_hel}
\alias{gmap_hel}
\title{Google Maps aerial image of HEL1}
\format{\code{raster::RasterBrick}}
\description{
This aerial image was downloaded from Google Maps and covers the 
research plot HEL1 declared within the framework of the research unit 
"Kilimanjaro Ecosystems Under Global Change: Linking biodiversity, biotic 
interactions and biogeochemical ecosystem processes". It was used to conduct 
unsupervised image classification based on \code{\link{kmeans}} clustering.
}
\details{
For a more detailed description of the applied analysis, please refer to 
\url{https://www.researchgate.net/publication/281716190_Unsupervised_classification_of_Google_maps_imagery_in_R} 
(accessed on January 27, 2016).
}
