% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hclPalette.R
\name{hclPalette}
\alias{hclPalette}
\title{create a color palette using hcl color space}
\usage{
hclPalette(n, c = 60, l = 50, ...)
}
\arguments{
\item{n}{length of the palette (i.e. how many colors)}

\item{c}{chroma value for the palette}

\item{l}{luminance value for the palette}

\item{...}{additional arguments to be passed to \code{\link{hcl}}}
}
\description{
create a color palette using hcl color space
}
\author{
Tim Appelhans
}

