% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histmatchRaster.R
\name{histmatchRaster}
\alias{histmatchRaster}
\title{histogram matching for RasterLayers}
\usage{
histmatchRaster(x, y, ttab = FALSE, minval = NULL, maxval = NULL,
  by = NULL, ...)
}
\arguments{
\item{x}{RasterLayer to be adjusted}

\item{y}{target RasterLayer}

\item{ttab}{logical, whether to return the transformation table}

\item{minval}{lower bound of the possible range for transformation
(if \code{NULL}, the minimum of both layers)}

\item{maxval}{upper bound of the possible range for transformation
(if \code{NULL}, the maximum of both layers)}

\item{by}{step size used to build the new historgram
(if \code{NULL}, 1 for integer master layer, 0.01 for double master layer)}

\item{...}{additional arguments (mask) passed to \code{\link[landsat]{histmatch}}}
}
\value{
if \code{ttab = FALSE} a RasterLayer, if \code{ttab = TRUE} a list
with components
\code{recode} the trasnsformation table used to match the histograms
\code{newraster} the transformed RasterLayer
}
\description{
this function is a wrapper around \code{\link[landsat]{histmatch}} for
RasterLayers.
}
\examples{
library(latticeExtra)
tofix <- raster(matrix(rnorm(10000, 100, 3), nrow = 100, ncol = 100))
master <- raster(matrix(rnorm(10000, 130, 7), nrow = 100, ncol = 100))

raw_p <- densityplot(tofix[], xlim = c(80, 160), plot.points = FALSE) +
 as.layer(densityplot(master[], xlim = c(80, 160),
          plot.points = FALSE, col = "red"))

raw_p

fixed <- histmatchRaster(tofix, master)
raw_p + as.layer(densityplot(fixed[], xlim = c(80, 160),
                 plot.points = FALSE, col = "black", lty = 2))

}
\seealso{
\code{\link[landsat]{histmatch}}
}
\author{
Tim Appelhans
}
