% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioa.R
\name{ioa}
\alias{ioa}
\title{Index of Association}
\usage{
ioa(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
numeric between 0 and 1. 0 means the two vectors are always 
changing in opposite directions, 1 means they are always changing
in identical directions
}
\description{
This is a very basic index of association (IOA). It simply counts how often 
two vectors increase or decrease in parallel.
}
\examples{
a <- 1:10
b <- 11:20
c <- 110:101
set.seed(123)
d <- rnorm(100, 7, 1)
set.seed(234)
e <- rnorm(100, 200, 10)

ioa(a, b)
ioa(a, c)
ioa(d, e)

}
\references{
Matloff N. (2011), The Art of R Programming (and page 49 therein).
}
\author{
Norman Matloff
}
