% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsenal-package.R
\docType{data}
\name{kili12}
\alias{kili12}
\title{KiLi FOR1246 Habitats}
\format{a data frame with the following columns\cr
\cr
Habitat    : factor with levels ordered according to their mean elevation\cr
Elevation  : numeric\cr}
\description{
data frame of KiLi FOR1246 Habitats and respective elevations
}
\details{
Note, the Habitats are ordered according to their mean elevation
}

