% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kiliAerial.R
\name{kiliAerial}
\alias{kiliAerial}
\title{Retrieve an (aerial) image of Mt. Kilimanjaro}
\usage{
kiliAerial(projection = "+init=epsg:21037", type = "bing",
  template = NULL, rasterize = FALSE, ...)
}
\arguments{
\item{projection}{Character, defaults to "+init=epsg:21037". Output projection
of the retrieved OSM data, passed on to \code{\link{openproj}}.}

\item{type}{Character, defaults to "bing". Tile server from which to get the
map, see \code{\link{openmap}}.}

\item{template}{\code{Extent} object, or any object from which an \code{Extent}
object can be extracted. Defaults to the extent of the 8-km GIMMS grid used
in \strong{Detsch et al. (in preparation)}.}

\item{rasterize}{Logical, defaults to FALSE. Determines whether a multi-layer
\code{Raster*} object is returned.}

\item{...}{Further arguments passed on to \code{\link{openmap}}.}
}
\value{
An object of class \code{OpenStreetMap} or, if \code{rasterize = TRUE},
a \code{RasterStack} object.
}
\description{
Based on the functionalities offered by the R \bold{OpenStreetMap} package,
this function per default retrieves a Bing aerial image centered on Mt.
Kilimanjaro, Tanzania, that is aligned with the 8-km GIMMS grid covering the
region.
}
\examples{
library(ggplot2)

img <- kiliAerial(minNumTiles = 20)
autoplot(img)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{openmap}}, \code{\link{openproj}}
}

