% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticeCombineLayer.R
\name{latticeCombineLayer-deprecated}
\alias{latticeCombineLayer-deprecated}
\title{combine multiple lattice plots layerwise}
\usage{
latticeCombineLayer(trellis.list, ...)
}
\arguments{
\item{trellis.list}{a list containing lattice plot objects}

\item{...}{additional arguments passed to \code{\link{as.layer}}}
}
\value{
a single lattice plot object
}
\description{
this function combines multiple lattice plot objects drawing 
each as a layer on top of the previous plots. Note that the 
global plot settings (e.g. xlim, ylim, ...) is taken from the 
first object. This is particularly useful when looping over large amounts of data
using \code{\link{lapply}} (see examples).
}
\seealso{
\code{\link{Rsenal-deprecated}}
}
\author{
Tim Appelhans
}
\keyword{internal}
