% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonDist.R
\name{lonDist}
\alias{lonDist}
\title{Calculate distance between two longitudes}
\usage{
lonDist(lon_min, lon_max = NULL, lat, to_km = TRUE, ...)
}
\arguments{
\item{lon_min}{\code{numeric}. The starting longitude.}

\item{lon_max}{\code{numeric}. The (optional) end longitude. If not supplied, 
the argument automatically equals \code{lon_min + 1}.}

\item{lat}{\code{numeric}. The latitude for which to calculate longitude
distance.}

\item{to_km}{\code{logical}. Determines whether the calculated distance is 
returned in kilometers (default) or miles.}

\item{...}{Currently not in use.}
}
\value{
The linear distance between the starting and end longitude at the given
latitude or, if argument \code{lon_max} is not supplied, the length of 1° of 
longitude at the given latitude.
}
\description{
This function is intended to calculate the distance between two longitudes 
at a given latitude. If only one longitude location is supplied, the length 
of 1° of longitude is calculated at the given latitude.
}
\examples{
lonDist(lon_min = 0, lat = 0)

}
\author{
Florian Detsch
}

