% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualBuild.R
\name{manualBuild}
\alias{manualBuild}
\title{Build package manually}
\usage{
manualBuild(dsn = getwd(), document = TRUE, ...)
}
\arguments{
\item{dsn}{'character'. Target folder containing source files; defaults to 
the current working directory.}

\item{document}{'logical'. Determines whether or not to invoke 
\code{\link{roxygenize}} with default roclets.}

\item{...}{Further arguments passed on to \code{\link[devtools]{build}}.}
}
\description{
This function was specifically designed to build a package from local source 
files manually, i.e., without using the package building functionality 
offered e.g. by RStudio.
}
\examples{
\dontrun{
## when in a package directory, e.g. '~/satellite' 
manualBuild()
}

}
\author{
Florian Detsch
}
\seealso{
\code{\link{roxygenize}}, \code{\link[devtools]{build}},\code{\link{install.packages}}.
}

