% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modeInteger.R
\name{modeInteger}
\alias{modeInteger}
\title{Calculate modal value from integer vector}
\usage{
modeInteger(v)
}
\arguments{
\item{v}{Integer. The input vector.}
}
\value{
The \code{integer} modal value.
}
\description{
This function computes the modal value, i.e. the value that appears most
often, from an integer vector.
}
\examples{
set.seed(10)
x <- sample(1:10, 100, replace = TRUE)

modeInteger(x)
}
\author{
Tim Appelhans
}

