% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/number2binary.R
\name{number2binary}
\alias{number2binary}
\title{Convert an integer to fixed-length binary format}
\usage{
number2binary(number, no_bits, to_char = FALSE)
}
\arguments{
\item{number}{An \code{integer} to be converted.}

\item{no_bits}{\code{integer}. The desired length of the returned bit code.}

\item{to_char}{\code{logical}. If \code{TRUE}, the resulting bit code is
returned as merged string.}
}
\value{
A bit code in fixed-length \code{integer} format or, if \code{to_char}, a
single \code{character}.
}
\description{
This is a wrapper function around \code{\link{intToBin}} that returns a bit
code in fixed-length integer format. The code is based on an SO post by
Paul Hiemstra, see \url{https://stackoverflow.com/questions/12088080/how-to-convert-number-into-binary-vector}.
}
\examples{
number2binary(10)
number2binary(10, no_bits = 4)
number2binary(10, no_bits = 4, to_char = TRUE)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{intToBin}}.
}

