% Generated by roxygen2 (4.0.2): do not edit by hand
\name{overlap}
\alias{overlap}
\title{Assess overlap in vectors}
\usage{
overlap(x1, x2, na.rm = F, wordy = T, x1name = NULL, x2name = NULL)
}
\arguments{
\item{x1}{vector of same class as \code{x2}}

\item{x2}{vector of same class as \code{x1}}

\item{na.rm}{logical, remove NAs from analysis}

\item{wordy}{logical, prints convenience information}
}
\value{
summary data.frame
}
\description{
Useful for learning the overlap of absolute and unique values between two vectors of the same type.  They
do not have to be the same length, and are allowed to contain NAs.
}
\examples{
x1 <- sample(1:50, 40, replace=T)
x2 <- sample(c(1:60, rep(NA, 30)), 55, replace=T)
overlap(x1, x2, na.rm=T)
overlap(x1, x2)
overlap(mtcars$gear, mtcars$cyl, x1name='gear', x2name='cyl')
}

