% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.extent.R
\name{panel.extent}
\alias{panel.extent}
\title{panel function for raster extent objects}
\usage{
panel.extent(ext, ...)
}
\arguments{
\item{ext}{extent object}

\item{...}{additional arguments passed on to \code{\link{panel.polygon}}}
}
\value{
A \code{\link{panel.polygon}} object
}
\description{
this function produces a \code{\link{panel.polygon}} from a 
\code{\link{extent}} object. This can then be used to visualise 
the extent on top of an \code{\link{spplot}} object.
}
\examples{
library(latticeExtra)
rst <- raster(ncols = 360, nrows = 180)
rst[] <- rnorm(360 * 180)

ex <- extent(c(27.5, 62.5, -17.5, 17.5))

spplot(rst) + layer(panel.extent(ex))

}
\author{
Tim Appelhans
}

