% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partialDep}
\alias{partialDep}
\title{Partial Dependence (single variable)}
\usage{
partialDep(model, df, xvar, n = 10, target.class = "1", ci = c(0.9, 0.5,
  0.3))
}
\arguments{
\item{model}{model object used to generate predictions.  Currently only built and tested for random forest.}

\item{df}{data.frame or data.table used to generate predictions with \code{model}}

\item{xvar}{character of length one; the x variable in \code{df} to assess for partial dependence with the response variable from \code{model}.}

\item{n}{numeric of length one; number of values between the min and max of \code{xvar} to score the model.
Note: this number is also how many replicates of \code{df} must be created and stored in memory.  Default is 10.}

\item{target.class}{character: Which category (class) of the target variable to use for predictions}

\item{ci}{numeric: specify any confidence intervals around the median response.}
}
\value{
data.table of output.  \code{cnt} refers to how many obs from \code{df} are within the fixed-width bin specified by \code{xvar}.
}
\description{
Calculate the partial dependence of a predictor variable on the response variable from a random forest classification model.
Rather than sequence through values of the predictor variable of interest and keep the other predictors at their median, this partial dependence
technique creates replicates of the entire dataset for each level of the x variable of interest from it's min to max.  This gives a more
realistic idea of the magnitude and direction of the x variable on the response.
}
\examples{
library('randomForest')
library('data.table')
DF <- mtcars
DF$vs <- factor(DF$vs)
rf <- randomForest(vs~mpg+cyl+drat+qsec+disp+gear+carb+hp, DF, ntrees=100)
pd <- partialDep(model=rf, df=DF, xvar='mpg')
pd[ci==0.5,] # median of response when sequenced through 'mpg'

## Plotting
plot(pd[cilev==0, xvar], pd[cilev==0, pred], type='l', ylim=c(0,1))
lines(pd[ci==.95, xvar], pd[ci==.95, pred], type='l', col='red')
lines(pd[ci==.05, xvar], pd[ci==.05, pred], type='l', col='green')
}
\seealso{
\code{\link{partialDepPlot}}
}

