% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partialDepAll}
\alias{partialDepAll}
\title{Partial Dependence Spark Lines (multiple variables)}
\usage{
partialDepAll(model, df, n = 10, xvars = NULL, target.class = "1",
  ci = c(0.75), plot = T, data = T, plot.yaxis.fixed = T, wordy = T)
}
\arguments{
\item{model}{model object used to generate predictions.  Currently only built and tested for random forest.}

\item{df}{data.frame or data.table used to generate predictions with \code{model}}

\item{n}{numeric of length one; number of values between the min and max of \code{xvar} to score the model.
Note: this number is also how many replicates of \code{df} must be created and stored in memory.  Default is 10.}

\item{xvars}{character vector; the x variables in \code{df} to assess for partial dependence with the response variable from \code{model}.  Defaults to choosing all variables from model.}

\item{target.class}{character: Which category (class) of the target variable to use for predictions}

\item{ci}{numeric: specify any confidence intervals around the median response.}

\item{plot}{logical: plot sparklines (\code{TRUE}), or no (\code{FALSE})}

\item{data}{logical: return summary table of output. yes (\code{TRUE}), or no (\code{FALSE})}

\item{plot.yaxis.fixed}{logical: \code{TRUE} uses same yaxis for all x variables (biggest max and smallest min across all variables).  code{\FALSE} uses variable specific axes.}

\item{wordy}{logical: print progress to the console}
}
\value{
list of output and plot
}
\description{
Calculate the partial dependence of a predictor variable on the response variable from a random forest classification model.
Rather than sequence through values of the predictor variable of interest and keep the other predictors at their median, this partial dependence
technique creates replicates of the entire dataset for each level of the x variable of interest from it's min to max.  This gives a more
realistic idea of the magnitude and direction of the x variable on the response.
}
\examples{
library('randomForest')
DF <- mtcars
DF$vs <- factor(DF$vs)
rf <- randomForest(vs~mpg+cyl+drat+qsec+disp+gear+carb+hp, DF, ntrees=100)
pda <- partialDepAll(model=rf, df=DF, n=10)
}
\seealso{
\code{\link{partialDep}}
}

