% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfCloud.R
\name{pdfCloud}
\alias{pdfCloud}
\title{create a word cloud from a bunch of pdfs}
\usage{
pdfCloud(pdf.path, exclude = NULL, nwords = 180, min.char.length = 5,
  max.char.length = 30, clrs = c("#8c510a", "#bf812d", "#35978f", "#01665e",
  "#7fbc41", "#4d9221"), seed = 123, wmin = NULL, ...)
}
\arguments{
\item{pdf.path}{path to the pdf folder}

\item{exclude}{optional path to a .txt file with words to be excluded 
from the cloud}

\item{nwords}{the number of words to be plotted}

\item{min.char.length}{minimum length of words to be included in the cloud}

\item{max.char.length}{maximum length of words to be included in the cloud}

\item{clrs}{vector of colors to be used for the word cloud}

\item{seed}{seed to be used for the random word placement}

\item{wmin}{see \code{\link{tagcloud}} for details}

\item{...}{additional arguments passed to \code{\link{tagcloud}}}
}
\description{
this function creates a wordcloud by converting a bunch of pdf files to
txt files. Optionally, a file with common words to be excluded can be 
supplied. See \code{\link{tagcloud}} for details on the cloud building.
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{tagcloud}}
}

