% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgBy.R
\name{plotAvgBy}
\alias{plotAvgBy}
\title{Plot average of an indicator over bins/categories of another another continuous variable}
\usage{
plotAvgBy(indv, byv, nbins = 5, data = F, plotNbin = T, ...)
}
\arguments{
\item{indv}{vector.  This the variable whose mean will be calculated over the categories of the \code{byv} vector binned up}

\item{byv}{vector. This is variable to be binned up, by which the \code{indv} variable will be averaged}

\item{nbins}{numeric. Number of bins to create when discretizing \code{byv}.  Passed to \code{depthbin} function.}

\item{data}{logical. TRUE returns the aggregated data.table. FALSE returns noting.  TRUE is default.}

\item{plotNbin}{logical.  TRUE plots the count of obs in each bin on top of each bar.  TRUE is default.}

\item{...}{additional barplot arguments.}
}
\value{
prints a barplot unless data==T, in which case the aggregated data.table is returned
}
\description{
Useful for assessing the relationship between two variables of interest.  There are cases, especially when 
outliers are involved, or many obs, that a scatterplot can be difficult to read.  This function bins up one of the continuous
variables that would be used in a scatterplot and calculates the mean (or other function) of the continuous variable over 
a range (discretized into categories) of the second continuous variable.  It uses an equal depth binning algorithm
to compute these bins on the \code{by} variable.  

It can be used to assess the average of a binary target variable/prediction
over a range of levels of a continuous or categorical variable.
}
\examples{

plotAvgBy(mtcars[,'mpg'], mtcars[,'drat'], nbins=8)
plotAvgBy(mtcars[,'mpg'], mtcars[,'drat'], nbins=5, plotNbin=F)
plotAvgBy(mtcars[,'mpg'], mtcars[,'drat'], nbins=5, plotNbin=F, data=T)

## Example with missing data
df <- mtcars
df$mpg[sample(1:nrow(mtcars), 5)] <- NA
df$drat[sample(1:nrow(mtcars), 5)] <- NA

plotAvgBy(df[,'mpg'], df[,'drat'], nbins=5, plotNbin=F, data=T)
}

