% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdf.R
\name{plotdf}
\alias{plotdf}
\title{Plot data.frame to PDF}
\usage{
plotdf(df, file = "output.pdf", wordy = F)
}
\arguments{
\item{df}{data.frame to plot.}

\item{file}{string - name of the PDF that will be created}

\item{wordy}{- boolean \code{TRUE} or \code{FALSE}.  Sequentially prints status of each chart to the console.
Could be useful for large data.frames.}
}
\value{
PDF file of plots
}
\description{
Plots every column of a data.frame as an individual plot (one plot per page)
 in a PDF file.
}
\examples{
\dontrun{
plotdf(df=mtcars, file='mtcars_plots.pdf')
}
}

