% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRfeCV.R
\name{plotRfeCV}
\alias{plotRfeCV}
\title{plot cross-validated variable importance}
\usage{
plotRfeCV(rfe.model, metric = rfe.model$metric, xlim = "minmax",
  ylim = "minmax")
}
\arguments{
\item{rfe.model}{a rfe model. See \code{\link{rfe}}}

\item{metric}{the metric to be used. Note this needs to be the metric used
to calculate the \code{\link{rfe}} model}

\item{xlim}{the xlim for the plot}

\item{ylim}{the ylim for the plot}
}
\value{
a trellis object
}
\description{
this function plots variable imporatnce for cross-validated
\code{\link{rfe}} variable selection classes. It uses the k cross-validations
to compute the mean +/- sd error metric.
}
\author{
Hanna Meyer, Tim Appelhans
}

