% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pos2just.R
\name{pos2just}
\alias{pos2just}
\title{Convert integer text position specifier to string}
\usage{
pos2just(pos, ...)
}
\arguments{
\item{pos}{Integer. A position specifier for text annotations as used by
\code{\link{text}}.}

\item{...}{Further arguments. Currently not in use.}
}
\value{
A character vector used as input for text justification in \strong{grid}-based text
drawing functions.
}
\description{
Convert integer position specifiers as supported by
\code{\link{text}} to character position specifiers as supported by
\strong{grid}-based text drawing functions (e.g. \code{\link{grid.text}}).
}
\examples{
set.seed(100)
pos <- sample(1:4, 5, replace = TRUE)

pos2just(pos)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{text}}, \code{\link{grid.text}}
}

