% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predictionStats.R
\name{predictionStats}
\alias{predictionStats}
\title{calculate prediction performance statistics}
\usage{
predictionStats(prd, obs, adj.rsq = TRUE, plot = FALSE)
}
\arguments{
\item{prd}{numeric vector of predicted values}

\item{obs}{numeric vector of observed values}

\item{adj.rsq}{logical, whether to return adjusted r-squared.
Defaults to TRUE}

\item{plot}{logical, whether to produce a visualisation of the results.
Defaults to FALSE}
}
\value{
If \code{plot = FALSE} (the default), a data frame.
If \code{plot = TRUE}, a list with components \code{stats} - data frame
and \code{plot} - a trellis plot object.
}
\description{
this function calculates prediction performance statistics
between vectors of predicted and observed values, namely
coefficient of determination (Rsq), root mean squared error (RMSE),
mean error (ME), mean absolute error (MAE). Users may also create
a dotplot visualising the results.
}
\examples{
pred_vals <- rnorm(1000, 200, 10)
obs_vals <- rnorm(1000, 200, 10)

result <- predictionStats(sort(pred_vals), sort(obs_vals), plot = TRUE)
result$stats
result$plot
}
\author{
Tim Appelhans
}

