% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts2spy.R
\name{pts2spy}
\alias{pts2spy}
\title{Convert points to SpatialPolygons*}
\usage{
pts2spy(coords, hole = NA, ID, data, match.ID = TRUE, ...)
}
\arguments{
\item{coords}{2-column \code{numeric} matrix with x and y coordinates.}

\item{hole}{\code{logical}, see \code{\link{Polygon}}.}

\item{ID}{\code{character}, see \code{\link{Polygons}}.}

\item{data}{\code{data.frame} with data to add to the output 
\code{SpatialPolygons*} object (optional).}

\item{match.ID}{\code{logical}, see \code{\link{SpatialPolygonsDataFrame}}.}
}
\value{
If \code{data} is missing, a \code{SpatialPolygons} object; else a 
\code{SpatialPolygonsDataFrame} object.
}
\description{
Create a \code{SpatialPolygons*} object from a set of point coordinates in 
one go, i.e. without being required to run through the single steps outlined 
in \code{\link{SpatialPolygons}}.
}
\examples{
coords1 <- cbind(c(2, 4, 4, 1, 2), c(2, 3, 5, 4, 2))
spy1 <- pts2spy(coords1, ID = "A")

coords2 <- cbind(c(5, 4, 2, 5), c(2, 3, 2, 2))
spy2 <- pts2spy(coords2, ID = "B")

plot(spy1, col = "grey75")
plot(spy2, col = "grey25", add = TRUE)

}
\author{
Florian Detsch
}
\seealso{
\code{\link{SpatialPolygons-class}}, \code{\link{SpatialPolygonsDataFrame}}.
}

