% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quantile_df}
\alias{quantile_df}
\title{Pretty print quantiles}
\usage{
quantile_df(x, probs, na.rm = F, names = F, type = 7, colname = NULL,
  ...)
}
\arguments{
\item{x}{numeric vector to calculate quantile with}

\item{probs}{numeric vector of probabilities to use for quantile}

\item{na.rm}{logical, pass to quantile function}

\item{names}{logical, pass to quantile function}

\item{type,}{number, pass to quantile function}

\item{colname,}{character, name of variable column in output}
}
\value{
prettified data.frame of probabilities
}
\description{
Useful for R Markdown reports
}
\examples{
quantile_df(mtcars$mpg, seq(0,1,.1))
quantile_df(mtcars$mpg, seq(0,1,.1), names=T)
quantile_df(mtcars$mpg, seq(0,1,.2), colname='mpg')
}

