% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbowPalette.R
\name{rainbowPalette}
\alias{rainbowPalette}
\title{Perceptually optimised rainbow colors}
\usage{
rainbowPalette(n)
}
\arguments{
\item{n}{the number of colours to be returned}
}
\value{
a vector of colours
}
\description{
This function returns a perceptually optimised rainbow color palette.
}
\examples{
pal <- function(col, border = "transparent", ...) {
  n <- length(col)
  plot(0, 0, type="n", xlim = c(0, 1), ylim = c(0, 1),
       axes = FALSE, xlab = "", ylab = "", ...)
  rect(0:(n-1)/n, 0, 1:n/n, 1, col = col, border = border)
} # courtesy Paul Murrell

pal(rainbowPalette(1000))

}
