% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rasterizeGimms.R
\name{rasterizeGimms}
\alias{rasterizeGimms}
\title{Rasterize GIMMS 3G binary data}
\usage{
rasterizeGimms(file, headerfile, water2na = TRUE, nodata2na = TRUE,
  scaling = TRUE, file_out = "", ...)
}
\arguments{
\item{file}{Character. GIMMS binary file to rasterize.}

\item{headerfile}{Character. Companion header file passed on to
\code{\link{read.ENVI}}. Usually created from \code{\link{createHdr}}.}

\item{water2na}{Logical. If TRUE, pixels holding 'water-mask' value (-10000)
are set to NA.}

\item{nodata2na}{Logical. If TRUE, pixels holding 'nodata-mask' value (-5000)
are set to NA.}

\item{scaling}{Logical. If TRUE, initial values are multiplied by scaling
factor (1/10000).}

\item{file_out}{Character. If supplied, \code{\link{writeRaster}} is invoked
to save the rasterized GIMMS data.}

\item{...}{Further arguments passed on to \code{\link{writeRaster}}.}
}
\value{
A \code{RasterLayer} object.
}
\description{
Convert GIMMS 3G binary data to an object of class \code{raster} by the use
of an ENVI header file.
}
\examples{
# Download GIMMS 3G data from 1990 to 2009 (this might take some time...)
fls_gimms <- downloadGimms(decades = c(1990, 2000),
                           dsn = paste(getwd(), "data", sep = "/"))

# Create companion header file
fls_hdr <- createHdr()

# Rasterize selected GIMMS data
rst_gimms <- rasterizeGimms(file = fls_gimms[1],
                            headerfile = fls_hdr)

plot(rst_gimms)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{createHdr}}, \code{\link{read.ENVI}}
}

