% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeIMERG.R
\name{rasterizeIMERG}
\alias{rasterizeIMERG}
\title{Rasterize GPM IMERG data}
\usage{
rasterizeIMERG(x, layer = "precipitationCal")
}
\arguments{
\item{x}{Filename of the IMERG HDF5 file}

\item{layer}{IMERG layer name}
}
\value{
A \code{RasterLayer}
}
\description{

}
\author{
Hanna Meyer
}

