% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeTRMM.R
\name{rasterizeTRMM}
\alias{rasterizeTRMM}
\title{Rasterize TRMM 3B42 Data}
\usage{
rasterizeTRMM(x, ...)
}
\arguments{
\item{x}{\code{character}. TRMM 3B42 file(s). Currently supported filetypes 
are .bin and .HDF. Note that \code{\link{rasterizeTRMMbin}} takes one file 
only.}

\item{...}{Additional arguments passed to \code{\link{rasterizeTRMMbin}} or 
\code{\link{rasterizeTRMMhdf}}.}
}
\value{
\itemize{
\item{\strong{.bin files}: A single \code{RasterLayer} object.}
\item{\strong{.HDF files}: A \code{list} of \code{Raster*} objects.}
}
}
\description{
Rasterize TRMM 3B42 .bin or .HDF files.
}
\examples{
 
\dontrun{
## download 3-hourly TRMM 3B42 data from Jan 1, 2015
fls <- downloadTRMM(begin = "2015-01-01", end = "2015-01-01", 
                    type = "3-hourly")
fls[1]                     

## rasterize
library(rworldmap)
eth <- subset(countriesCoarse, ADMIN == "Ethiopia")
rst <- rasterizeTRMM(fls, ext = eth)

## visualize
plot(rst[[1]])
}

}
\author{
Florian Detsch
}
\seealso{
\code{\link{downloadTRMM}}, \code{\link{readBin}}, 
\code{\link{rasterizeTRMMbin}}, \code{\link{rasterizeTRMMhdf}}.
}

