% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeTRMMbin.R
\name{rasterizeTRMMbin}
\alias{rasterizeTRMMbin}
\title{Rasterize TRMM 3B42 Binary Data}
\usage{
rasterizeTRMMbin(x, meta, write_out = TRUE)
}
\arguments{
\item{x}{\code{character}. TRMM 3B42 endian binary (.bin) file.}

\item{meta}{\code{character}. Metadata file (.xml) associated with 'x'.}

\item{write_out}{\code{logical}. Enables output storage.}
}
\value{
A \code{RasterLayer} object.
}
\description{
Rasterize a single TRMM 3B42 binary file (.bin).
}
\examples{
 
## see ?rasterizeTRMM

}
\author{
Florian Detsch
}
\references{
GES DISC (2010) Readme for TRMM Product 3B42 (v6). Available online: 
\url{ftp://ftp.iap.ac.cn/ftp/ds107_TRMM-3B42_0.25_3hour_hdf/TRMM_3B42_readme.htm}.
}
\seealso{
\code{\link{readBin}}, \code{\link{downloadTRMM}}.
}

