% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeTRMMhdf.R
\name{rasterizeTRMMhdf}
\alias{rasterizeTRMMhdf}
\title{Rasterize TRMM 3B42 HDF Data}
\usage{
rasterizeTRMMhdf(x, SDSstring = "111111", ext = NULL, cores = 1L,
  filename = "", ...)
}
\arguments{
\item{x}{\code{character}. TRMM 3B42 .HDF file(s).}

\item{SDSstring}{\code{character}. Determines which SDS layers to process 
(1 = process, 0 = do not process). The default is to process all SDS layers 
(i.e., "111111").}

\item{ext}{\code{Extent}, or any object from which an \code{Extent} can be 
extracted, see \code{\link[raster]{crop}}.}

\item{cores}{\code{integer}. Number of cores for parallel processing.}

\item{filename}{\code{character}. Optional output filename(s). If specified, 
this must be of the same length as 'x'.}

\item{...}{Additional arguments passed to \code{\link{writeRaster}}.}
}
\value{
A \code{list} of \code{Raster*} objects, with each entry corresponding to 
one element in 'x' and the number of layers per \code{Raster*} object 
corresponding to the processed SDS layers.
}
\description{
Rasterize TRMM 3B42 HDF files.
}
\examples{
## see ?rasterizeTRMM

}
\references{
GES DISC (2016) Readme for TRMM Product 3B42 (V7). Available online: 
\url{http://disc.sci.gsfc.nasa.gov/precipitation/documentation/TRMM_README/TRMM_3B42_readme.shtml}.
}

