% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rearrangeGimms.R
\name{rearrangeGimms}
\alias{rearrangeGimms}
\title{Rearrange GIMMS 3G binary files}
\usage{
rearrangeGimms(fls = NULL, dsn = ".", pattern = "^geo",
  rename_yearmon = FALSE, ...)
}
\arguments{
\item{fls}{Character. Vector of filepaths. If \code{NULL}, folder \code{dsn}
will be searched for available GIMMS data to rearrange.}

\item{dsn}{Character. Path to look for GIMMS data. If not supplied and
\code{is.null(fls)}, this argument defaults to the current working directory.}

\item{pattern}{Character. A regular expression passed on to
\code{\link{list.files}}.}

\item{rename_yearmon}{Logical. If TRUE, date information in GIMMS filenames
is replaced by a more intuitive format that facilitates file sorting (e.g.,
"81jul" is replaced by "198107").}

\item{...}{Further arguments passed on to \code{\link{list.files}}.}
}
\value{
A vector of GIMMS filepaths in ascending order of time.
}
\description{
Rearrange (previously downloaded) GIMMS 3G binary data in ascending order of
time.
}
\examples{
# Download GIMMS 3G data from 1990 to 2009 (this might take some time...)
gimms_fls <- downloadGimms(decades = c(1990, 2000),
                           dsn = paste(getwd(), "data", sep = "/"))

# Rearrange downloaded GIMMS data in ascending order of time
rearrangeGimms(fls = gimms_fls)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{list.files}}
}

