% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionStats.R
\name{regressionStats}
\alias{regressionStats}
\title{calculate prediction performance statistics}
\usage{
regressionStats(prd, obs, adj.rsq = TRUE, plot = FALSE,
  method = "pearson")
}
\arguments{
\item{prd}{numeric vector of predicted values}

\item{obs}{numeric vector of observed values}

\item{adj.rsq}{logical, whether to return adjusted r-squared. 
Defaults to TRUE}

\item{plot}{logical, whether to produce a visualisation of the results.
Defaults to FALSE}

\item{method}{character. Method to use for correlation. See \code{?cor.test}
for details.}
}
\value{
If \code{plot = FALSE} (the default), a data frame. 
If \code{plot = TRUE}, a list with components \code{stats} - data frame
and \code{plot} - a trellis plot object.
}
\description{
this function calculates prediction performance statistics
between vectors of predicted and observed values, namely
coefficient of determination (Rsq), root mean squared error (RMSE), 
mean error (ME), mean absolute error (MAE). Users may also create 
a dotplot visualising the results.
}
\examples{
## create predictions with high accuracy (identical mean), 
## but low precision (sd double that of observations). Hence,
## ME should be close to zero and RMSE close to ten.
pred_vals <- sort(rnorm(1000, 200, 20)) # sorting ensures high Rsq
obs_vals <- sort(rnorm(1000, 200, 10))

## with plot = TRUE
result <- regressionStats(pred_vals, obs_vals, plot = TRUE)
result$stats
result$plot

## with plot = FALSE
result <- regressionStats(pred_vals, obs_vals, 
                          plot = FALSE, adj.rsq = FALSE)
result

}
\author{
Tim Appelhans, Hanna Meyer
}
