% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeNA.R
\name{removeNA}
\alias{removeNA}
\title{remove rows with NA values in selcted columns of a data frame}
\usage{
removeNA(df, cols)
}
\description{
remove rows with NA values in selcted columns of a data frame
}

