% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rfImp}
\alias{rfImp}
\title{Sort Random Forest Variable Importance Features}
\usage{
rfImp(rf)
}
\arguments{
\item{rf}{randomForest object from randomForest package}
}
\value{
data.frame
}
\description{
Simple utility function to sort random forest variable importance features.  Very simple.
}
\examples{
library('randomForest')
myrf <- randomForest(iris[1:4], iris$Species)
rfImp(myrf)
}

