% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2spLayout.R
\name{rgb2spLayout-deprecated}
\alias{rgb2spLayout-deprecated}
\title{Convert an RGB RasterBrick/Stack to use with spplot}
\usage{
rgb2spLayout(x, quantiles = c(0.02,0.98), alpha = 1)
}
\arguments{
\item{x}{a RasterBrick or RasterStack}

\item{quantiles}{Upper and lower quantiles used for color stretching.}

\item{alpha}{Level of transparency.}
}
\description{
This function takes a red-green-blue 'RasterStack' (or 'RasterBrick') object
and produces a list with color information that can be passed on to the
\code{sp.layout} argument from \code{\link[sp]{spplot}}.
}
\seealso{
\code{\link{Rsenal-deprecated}}
}
\author{
Tim Appelhans, Florian Detsch
}
\keyword{internal}
