% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2YCbCr.R
\name{rgb2YCbCr}
\alias{rgb2YCbCr}
\title{Convert from RGB to YCbCr color space}
\usage{
rgb2YCbCr(x, r = 1, g = 2, b = 3)
}
\arguments{
\item{x}{A \code{RasterStack} or \code{RasterBrick} object. 3
bands are mandatory (usually red, green and blue).}

\item{r}{Integer, defaults to '1'. Index of the red channel.}

\item{g}{Integer, defaults to '2'. Index of the green channel.}

\item{b}{Integer, defaults to '3'. Index of the blue channel.}
}
\value{
A true-color \code{RasterStack} object converted to YCbCr color space.
}
\description{
This function transforms a RGB \code{Raster*} object to YCbCr color space.
}
\examples{
b_rgb <- brick(system.file("external/rlogo.grd", package="raster"))
plotRGB(b_rgb)

b_ycbcr <- rgb2YCbCr(b_rgb)
plot(b_ycbcr, 
     main = c("luminance", "blue-yellow chrominance", "green-red chrominance"))

}
\references{
Deb, K. and Suny, A.H. (2014): Shadow Detection and Removal Based on YCbCr Color Space. Smart Computing Review 4, 23-33, doi:10.6029/smartcr.2014.01.003.
}
\author{
Florian Detsch, Tim Appelhans
}
