% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgbShadowMask.R
\name{rgbShadowMask}
\alias{rgbShadowMask}
\title{Shadow mask from RGB (satellite) imagery}
\usage{
rgbShadowMask(rgb, n_iter = 1, ...)
}
\arguments{
\item{rgb}{A \code{RasterStack} or \code{RasterBrick} object. 3
bands are mandatory (usually red, green and blue).}

\item{n_iter}{Integer, defaults to '1'. Number of iterations to distinguish
between shadow and non-shadow pixels. The more iterations, the fewer shadow
pixels will remain.}

\item{...}{Further arguments passed on to \code{\link{rgb2YCbCr}}.}
}
\value{
A shadow mask \code{RasterLayer}.
}
\description{
This function builds a shadow mask from a RGB \code{Raster*} object.
}
\examples{
library(RColorBrewer)

data(gmap_hel)
plotRGB(gmap_hel)

gmap_hel_wsh <- rgbShadowMask(gmap_hel)
plot(gmap_hel_wsh, axes = FALSE, legend = FALSE,
     col = c(rev(brewer.pal(9, "Greys"))), alpha = .25, add = TRUE)
}
\author{
Florian Detsch, Tim Appelhans
}
\references{
Deb, K. and Suny, A.H. (2014): Shadow Detection and Removal Based on YCbCr Color Space. Smart Computing Review 4, 23-33, doi:10.6029/smartcr.2014.01.003.
}
\seealso{
\code{\link{rgb2YCbCr}}
}

