% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsenal-deprecated.R,
%   R/bumpVersion.R, R/calcOffsetGridText.R, R/compareDistributions.R,
%   R/deprecated_CreateSpacetimeFolds.R, R/deprecated_ffs.R,
%   R/latticeCombineGrid.R, R/latticeCombineLayer.R, R/offsetGridText.R,
%   R/pos2just.R, R/rgb2spLayout.R, R/stextGrob.R
\name{Rsenal-deprecated}
\alias{Rsenal-deprecated}
\alias{bumpVersion}
\alias{calcOffsetGridText}
\alias{compareDistributions}
\alias{CreateSpacetimeFolds}
\alias{ffs}
\alias{latticeCombineGrid}
\alias{latticeCombineLayer}
\alias{offsetGridText}
\alias{pos2just}
\alias{rgb2spLayout}
\alias{stextGrob}
\title{Deprecated functions in \pkg{Rsenal}}
\usage{
bumpVersion(element = "patch", pkg.repo = ".", news = file.path(pkg.repo,
  "NEWS.md"), plain_news = TRUE)

calcOffsetGridText(x, y = NULL, offset = 0.02, pos = NULL, xlim = NULL,
  ylim = NULL)

compareDistributions(left, right, add.spread = TRUE, print.stats = TRUE,
  xlim = NULL, ylim = NULL, clrs = c("purple", "lightblue"),
  xlab = "density", ylab = "value", ...)

CreateSpacetimeFolds(x, spacevar = NA, timevar = NA, k = 10,
  seed = sample(1:1000, 1))

ffs(predictors, response, method = "rf",
  metric = ifelse(is.factor(response), "Accuracy", "RMSE"),
  maximize = ifelse(metric == "RMSE", FALSE, TRUE), withinSD = FALSE,
  trControl = trainControl(), tuneLength = 3, tuneGrid = NULL,
  seed = sample(1:1000, 1), runParallel = FALSE, ...)

latticeCombineGrid(trellis.list, between = list(y = 0.3, x = 0.3),
  as.table = TRUE, ...)

latticeCombineLayer(trellis.list, ...)

offsetGridText(x, y = NULL, labels, xlim = NULL, ylim = NULL,
  pos = NULL, stext = FALSE, offset = 0.02, ...)

pos2just(pos, ...)

rgb2spLayout(x, quantiles = c(0.02, 0.98), alpha = 1)

stextGrob(label, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  col = "white", fill = "black", r = 0.1, gp = gpar(), vp = NULL,
  name = NULL, ...)
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("Rsenal-deprecated")}.

\itemize{
  \item \code{\link[Rsenal]{bumpVersion-deprecated}}
  \item \code{\link[Rsenal]{calcOffsetGridText-deprecated}}
  \item \code{\link[Rsenal]{compareDistributions-deprecated}}
  \item \code{\link[Rsenal]{CreateSpacetimeFolds-deprecated}}
  \item \code{\link[Rsenal]{ffs-deprecated}}
  \item \code{\link[Rsenal]{latticeCombineGrid-deprecated}}
  \item \code{\link[Rsenal]{latticeCombineLayer-deprecated}}
  \item \code{\link[Rsenal]{offsetGridText-deprecated}}
  \item \code{\link[Rsenal]{pos2just-deprecated}}
  \item \code{\link[Rsenal]{rgb2spLayout-deprecated}}
  \item \code{\link[Rsenal]{stextGrob-deprecated}}
}
}
\section{\code{bumpVersion}}{

For \code{bumpVersion}, use \code{\link[Orcs]{bumpVersion}} instead.
}

\section{\code{calcOffsetGridText}}{

For \code{calcOffsetGridText}, use \code{\link[Orcs]{offsetGridText}} and 
non-exported \code{Orcs:::calcOffsetGridText} instead.
}

\section{\code{compareDistributions}}{

For \code{compareDistributions}, use \code{\link[Orcs]{compareDistributions}} 
instead.
}

\section{\code{CreateSpacetimeFolds}}{

\code{CreateSpacetimeFolds} is deprecated (no further details available).
}

\section{\code{ffs}}{

\code{ffs} is deprecated (no further details available).
}

\section{\code{latticeCombineGrid}}{

For \code{latticeCombineGrid}, use \code{\link[Orcs]{latticeCombineGrid}} 
instead.
}

\section{\code{latticeCombineLayer}}{

For \code{latticeCombineLayer}, use \code{\link[Orcs]{latticeCombineLayer}} 
instead.
}

\section{\code{offsetGridText}}{

For \code{offsetGridText}, use \code{\link[Orcs]{offsetGridText}} instead.
}

\section{\code{pos2just}}{

For \code{pos2just}, use \code{Orcs:::pos2just} instead.
}

\section{\code{rgb2spLayout}}{

For \code{rgb2spLayout}, use \code{\link[Orcs]{rgb2spLayout}} instead.
}

\section{\code{stextGrob}}{

For \code{stextGrob}, use \code{\link[Orcs]{stextGrob}} instead.
}

\keyword{internal}
