\docType{package}
\name{Rsenal}
\alias{Rsenal}
\alias{Rsenal-package}
\title{a(R)senal of assorted R packages}
\description{
  This is mainly a collection of random, or more accurately, clusters of random R functions
  I've written for personal and work projects over the years.  I don't expect it to be immediately 
  useful to the general population of R users.  For now it's mostly my cloud storage (with nice documentation) for R functions
  that I use repeatedly.  Some functions are fairly general purpose, while others were developed for specific
  problems and are not super generalizeable in the current state. \cr 
  
  \bold{Current clusters of functionality:} \cr
  - manipulating igraph objects \cr
  - modeling and model validation \cr
  - cleaning data \cr
  - scaling and transforming data 
}

\author{Andrew Brooks \cr \email{andrewbrooksct@gmail.com}}


\details{
  \tabular{ll}{
  Package: \tab Rsenal\cr
  Type: \tab Package\cr
  Version: \tab 0.1\cr
  Date: \tab 2014-10-08\cr
  }
}