% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rules2df}
\alias{rules2df}
\title{Transform association rule results into data.frame}
\usage{
rules2df(rules, list = F)
}
\arguments{
\item{rules}{list of association rules (S4 arules object).  Output of \code{apriori} function.}

\item{list}{logical \code{TRUE} or \code{FALSE}.  sets LHS items in a list, rather than one character string}
}
\value{
association rules in a data.frame
}
\description{
Note this function only currently works when the itemsets are of size 1 on the LHS and RHS
}
\examples{
library('arules')
data("Adult")
ar <- apriori(Adult, parameter = list(supp = 0.5, conf = 0.6, target = "rules", minlen=2))
df <- rules2df(ar, list=T)
}

