% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{calculate the standard error of the mean}
\usage{
se(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
this function calculates the standard error of the mean.
}
\examples{
a <- rnorm(1000, 20, 5)
mean(a)
se(a)
}
\author{
Tim Appelhans
}

