% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/significantMannKendall.R
\name{significantMannKendall}
\alias{significantMannKendall}
\title{Man-Kendall test for significant trends}
\usage{
significantMannKendall(x, p = 0.001)
}
\arguments{
\item{x}{Numeric. A vector of data.}

\item{p}{Numeric. Significance level.}
}
\value{
A numeric or, if the return value is not sufficiently significant, \code{NA}.
}
\description{
In contrast to the ordinary \code{\link{MannKendall}} function included in
the \bold{Kendall} package, this function returns significant trends
(according to user-defined p value) only.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{MannKendall}}
}

