% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantTau.R
\name{significantTau}
\alias{significantTau}
\title{Compute Kendall's tau}
\usage{
significantTau(x, p = 0.001, prewhitening = FALSE, df = FALSE, ...)
}
\arguments{
\item{x}{A 'numeric' vector.}

\item{p}{Significance level to be tested.}

\item{prewhitening}{'logical'. If \code{TRUE}, pre-whitening (see description 
in package \strong{zyp}) is applied prior to the Mann-Kendall trend test.}

\item{df}{'logical'. If \code{TRUE}, a 'data.frame' holding the value of 
Kendall's tau and the referring significance level.}

\item{...}{Further arguments passed on to \code{\link{zyp.trend.vector}}.}
}
\value{
Kendall's tau depending on the predefined significance level. If 
\code{p} is exceeded, this automatically defaults to \code{NA}.
}
\description{
Apply the Mann-Kendall trend test (Mann, 1945) to a series of observations 
and return Kendall's tau (Kendall, 1937) based on a predefined significance 
level. In contrast to other readily available implementations, it is left to 
the user to decide whether or not to apply pre-whitening.
}
\examples{
## Example taken from ?Kendall::MannKendall
library(Kendall)
data(PrecipGL)
plot(PrecipGL)

## Mann-Kendall trend test without pre-whitening
significantTau(PrecipGL, p = 0.001)

## Mann-Kendall trend test with pre-whitening
significantTau(PrecipGL, p = 0.001, prewhitening = TRUE)

}
\author{
Florian Detsch
}
\seealso{
\code{\link{MannKendall}}, \code{\link{zyp.trend.vector}}.
}

