% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitRaster.R
\name{splitRaster}
\alias{splitRaster}
\title{Split a Raster object}
\usage{
splitRaster(x, s = 2, ...)
}
\arguments{
\item{x}{Character. File(s) to process.}

\item{s}{Integer. Number of slices 'x' should be divided into. Note that the 
default setting (\code{s = 2}) splits 'x' into four equally sized tiles, i.e. 
two in x-direction and two in y-direction.}

\item{...}{Further arguments passed on to \code{\link{gdal_translate}}.}
}
\value{
A list of RasterBrick objects.
}
\description{
Split one or multiple \code{Raster*} object(s) into a given number of 
horizontal and vertical slices. Inspired by 
\url{http://java.dzone.com/articles/use-gdal-r-console-split}.
}
\seealso{
\code{\link{gdal_translate}}, \code{\link{GDALinfo}}
}
\author{
Florian Detsch
}
