% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standardFocalWindow.R
\name{standardFocalWindow}
\alias{standardFocalWindow}
\title{create squared matrix for use with raster::focal}
\usage{
standardFocalWindow(n = 3, center = 0L)
}
\arguments{
\item{n}{size (i.e. ncol and nrow)}

\item{center}{the value (weight) for the center of the window. Defaults to 0}
}
\value{
A \code{\link{matrix}} of 1s with \code{n} columns and \code{n} rows and
a center value of \code{center}
}
\description{
this function produces a square matrix for use with \code{\link{focal}}.
User can set the size (must be odd) and center value (i.e. 0 or 1). All
non-center values are set to 1.
}
\examples{
standardFocalWindow() # 3 by 3 matrix
standardFocalWindow(n = 5, center = 1)
}
\author{
Tim Appelhans
}

