% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trainSeeds.R
\name{trainSeeds}
\alias{trainSeeds}
\title{create a list of seeds for trainControl()}
\usage{
trainSeeds(size, number, seed = 357)
}
\arguments{
\item{size}{the number of models to be evaluated}

\item{number}{the number of boot-starpping/cv repeats}

\item{seed}{the global seed to be set}
}
\value{
a list of integer seeds
}
\description{
this function creates a list of seeds to be used with \code{\link{trainControl}}
}
\examples{
trainSeeds(9, 3)
}
\author{
Tim Appelhans (largely taken from the examples of \code{\link{trainControl}})
}

