% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_network.R
\name{travOut}
\alias{travOut}
\title{Find all nodes connected to root node}
\usage{
travOut(g, root, orderN = -1)
}
\arguments{
\item{g}{igraph object}

\item{root}{character string: name of root node}

\item{orderN}{number: # of degrees away from root node to search.}
}
\value{
igraph object, subgragh of \code{g}
}
\description{
Returns a subset of the original graph (all edges and vertices that a
directed path can take from from the root node.  OrderN limits the growth of these paths.
}
\examples{
require('igraph')

## build sample network
from <- c('A','A','B','B','C','C','D','D','E','E','E','F','F','H','L','L','O')
to <- c('B','C','D','E','F','G','H','I','J','J','K','L','G','M','N','O','P')
relations<-cbind(from,to)
g <- graph.data.frame(relations)
tiers <- c(1,2,3,4,1,3,4,2,1,2,3,4,0,0,1,2)
V(g)$tier <- tiers

plot(g) ## full network 
plot(travOut(g, 'D')) ## sub network
}
\seealso{
Simpler version: \code{\link{travCount}}
}

