% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv2wdws.R
\name{uv2wdws}
\alias{uv2wdws}
\title{convert u and v components to wind direction and wind speed}
\usage{
uv2wdws(u, v)
}
\arguments{
\item{u}{a vector of u components}

\item{v}{a vector of v components}
}
\value{
a matrix
}
\description{
this function converts u and v components to wind direction and wind speed
}
\examples{
set.seed(123)
wd <- as.integer(rnorm(10, 180, 90))

set.seed(123)
ws <- rnorm(10, 4, 1)

## convert to u and v
wdws2uv(wd, ws)

## convert back
uv <- wdws2uv(wd, ws)
uv2wdws(uv[, 1], uv[, 2]) 

}
\author{
Tim Appelhans
}
\seealso{
\code{\link{wdws2uv}}
}

