% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/varsRfeCV.R
\name{varsRfeCV}
\alias{varsRfeCV}
\title{extract cross-validated important variables}
\usage{
varsRfeCV(rfe.model, metric = rfe.model$metric)
}
\arguments{
\item{rfe.model}{a rfe model. See \code{\link{rfe}}}

\item{metric}{the metric to be used. Note this needs to be the metric used
to calculate the \code{\link{rfe}} model}
}
\value{
a character vector of the variable names
}
\description{
this function approaches the identification of important variables from
\code{\link{rfe}} more conservatively than \code{\link{caret}}. It uses
the standard deviation of the cross-validated error metric to identify
important variables.
}
\author{
Hanna Meyer, Tim Appelhans
}

