% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPalette.R
\name{viewPalette}
\alias{viewPalette}
\title{view color ramps in a palette preview}
\usage{
viewPalette(col, border = "transparent", name = "")
}
\arguments{
\item{col}{the color palette to be visualised}

\item{border}{color of the segment borders. Defaults to "transparent"}

\item{name}{an optional name to be put into the plot}
}
\description{
This will create a plot of a color ramp
}
\examples{
library(colorspace)
library(viridisLite)

clrs <- list(rainbow = rainbow(100),
             rainbow_grey = desaturate(rainbow(100)),
             viridis = viridis(100),
             viridis_grey = desaturate(viridis(100)),
             inferno = inferno(100),
             inferno_grey = desaturate(inferno(100)),
             magma = magma(100),
             magma_grey = desaturate(magma(100)))

out <- lapply(seq(clrs), function(i) {
  viewPalette(clrs[[i]], name = names(clrs)[i])
})

latticeCombineGrid(out)

}
\author{
Tim Appelhans
}

