% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visDEM.R
\name{visDEM}
\alias{visDEM}
\title{Display Contour Lines}
\usage{
visDEM(dem, zlevs.conts = seq(1000, 5500, 500), labels = c(1000, "", 2000,
  "", 3000, "", 4000, "", 5000, ""), cex = 1.8, col = "grey50",
  labcex = 1, ...)
}
\arguments{
\item{dem}{DEM as \code{RasterLayer} or \code{character} file path.}

\item{zlevs.conts}{Elevation levels for which to include contour lines.}

\item{labels}{Labels associated with 'zlevs.conts'.}

\item{cex, col, labcex}{Additional graphical parameters, see 
\code{\link[graphics]{contour}} and \code{\link{par}}.}

\item{...}{Additional arguments passed to \code{Rsenal:::panel.smoothconts}.}
}
\value{
An object of class \code{"trellis"}.
}
\description{
Display contour lines deduced from a digital elevation model (DEM) in a way 
compatible with all kinds of \code{"trellis"} graphics.
}
\seealso{
\code{\link[graphics]{contour}}, \code{\link{Lattice}}.
}
\author{
Tim Appelhans, Florian Detsch
}
