% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vvi.R
\name{vvi}
\alias{vvi}
\title{Visible vegetation index}
\usage{
vvi(rgb, r = 1, g = 2, b = 3)
}
\arguments{
\item{rgb}{A \code{RasterStack} or \code{RasterBrick} object. 3
bands are mandatory (usually red, green and blue).}

\item{r}{Integer, defaults to '1'. Index of the red channel.}

\item{g}{Integer, defaults to '2'. Index of the green channel.}

\item{b}{Integer, defaults to '3'. Index of the blue channel.}
}
\value{
A VVI \code{RasterLayer}.
}
\description{
This function calculates a visible vegetation index (VVI) from a RGB
\code{Raster*} object.
}
\examples{
library(RColorBrewer)

data(gmap_hel)
plotRGB(gmap_hel)

gmap_hel_veg <- vvi(gmap_hel)
plot(gmap_hel_veg, col = brewer.pal(5, "BrBG"), alpha = .5, add = TRUE)
}
\author{
Florian Detsch, Tim Appelhans
}
\references{
Planetary Habitability Laboratory (2015): Visible Vegetation Index (VVI). Available online via \url{http://phl.upr.edu/projects/visible-vegetation-index-vvi}.
}

