% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borgIndices.R
\name{borgIndices}
\alias{borgIndices}
\title{Calculate shape indices listed in Borg 1998}
\usage{
borgIndices(Ar, Ur, De, Du)
}
\arguments{
\item{Ar}{area of the entities}

\item{Ur}{(requires documentation)}

\item{De}{Diameter of the inner circle}

\item{Du}{Diameter of the outer circle}
}
\value{
A RasterStack
}
\description{
used as internal function. see \code{?geometryVariables}
and Borg 1998: Vergleichende Analyse von Formindizes zur Charakterisierung 
von Landschaftsobjekten unter oekologischen Aspekten. 
Zeitschrift fuer Photogrammetrie und Fernerkundung 4: 108-119.
}
\author{
Hanna Meyer
}
\seealso{
\code{\link{geometryVariables}}.
}

