% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bumpVersion.R
\name{bumpVersion}
\alias{bumpVersion}
\title{bump package 'Version:' and 'Date:' in DESCRIPTION file}
\usage{
bumpVersion(element = "patch", pkg.repo = ".", news = file.path(pkg.repo,
  "NEWS"))
}
\arguments{
\item{element}{character - one of "major", "minor", "patch" to be bumped.
Defaults to "patch"}

\item{pkg.repo}{path to package repository folder. Default is current
working directory}

\item{news}{the NEWS file of the repo (assumed to be in top level path). 
If this exists, the first line of that file will be rewritten 
to be "<packagename> <major.minor.patch>"}
}
\description{
this function let's you bump the version number and creation date of
your package's DESCRIPTION file.
Supported versioning system is major.minor.patch
}
\author{
Tim Appelhans
}

