% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareDistributions.R
\name{compareDistributions}
\alias{compareDistributions}
\title{compare two density distributions side by side}
\usage{
compareDistributions(left, right, add.spread = TRUE, print.stats = TRUE,
  xlim = NULL, ylim = NULL, clrs = c("purple", "lightblue"),
  xlab = "density", ylab = "value", ...)
}
\arguments{
\item{left}{numeric vector}

\item{right}{numeric vector}

\item{clrs}{a character vector of length 2 specifying the colors
for the filled density regions}

\item{...}{additional arguments passed to \code{\link{density}}}

\item{add.spred}{logical, whether to plot the spread (q25 to q75 and the median)}
}
\value{
A trellis object
}
\description{
this function will produce a plot of two density functions displayed
side by side
}
\examples{
compareDistributions(rnorm(1000, 2, 3), rnorm(1000, -5, 1))
compareDistributions(rnorm(1000, 2, 3), rnorm(1000, -5, 1),
                     add.spread = FALSE)
compareDistributions(rnorm(1000, 2, 3), rnorm(1000, -5, 1),
                     add.spread = TRUE, clrs = c("red", "brown"))
}
\author{
Tim Appelhans
}

