% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr2Geos.R
\name{cr2Geos}
\alias{cr2Geos}
\title{column/row coordinates to geos projection}
\usage{
cr2Geos(x, COFF = 1856, ColumnDirGridStep = 3000.40316581726)
}
\arguments{
\item{x}{RasterLayer or RasterStack of the msg data where extent is defined in
column/row coordinates}

\item{COFF}{see description}

\item{ColumnDirGridStep}{see description}
}
\value{
RasterLayer or RasterStack in Geos projection
}
\description{
Converts column/row coordinates of MSG images to geos projection.
For HRV use COFF=5566 and ColumnDirGridStep=1000.13434886932 instead of the 
default values.
}
\examples{
cr2Geos(raster(extent(2200, 3000, 2450, 3050)))

}
\author{
Hanna Meyer and Johannes Droenner
}
