% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadGimms.R
\name{downloadGimms}
\alias{downloadGimms}
\title{Download GIMMS 3G data}
\usage{
downloadGimms(begin = NULL, end = NULL, dsn = ".", ...)
}
\arguments{
\item{begin}{Numeric. Start year, e.g. '2000'. If not supplied, download will
start from the first year available.}

\item{end}{Numeric. End year. If not supplied, download will stop at the last
year available.}

\item{dsn}{Character. Destination folder for file download. If not supplied,
all downloaded files will be stored in the current working directory.}

\item{...}{Further arguments. Currently not in use.}
}
\value{
A vector of filepaths.
}
\description{
Download GIMMS 3G binary data for a given time span from NASA FTP server
(\url{http://ecocast.arc.nasa.gov/data/pub/gimms/3g.v0/}).
}
\examples{
# Download GIMMS 3G data from 1990 to 2009 (this might take some time...)
downloadGimms(begin = 2000, dsn = paste(getwd(), "data", sep = "/"))
}
\author{
Florian Detsch
}
\seealso{
\code{\link{download.file}}
}

