% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTRMM.R
\docType{methods}
\name{downloadTRMM}
\alias{downloadTRMM}
\alias{downloadTRMM,POSIXlt-method}
\alias{downloadTRMM,character-method}
\title{Download TRMM 3B42 data}
\usage{
\S4method{downloadTRMM}{character}(begin, end, type = c("daily", "3-hourly"),
  dsn = getwd(), format = "\%Y-\%m-\%d")

\S4method{downloadTRMM}{POSIXlt}(begin, end, type = c("daily", "3-hourly"),
  dsn = getwd())
}
\arguments{
\item{begin}{\code{POSIXlt} or \code{character}, starting time.}

\item{end}{Same as \code{begin}, end time.}

\item{type}{\code{character}, data repetition cycle. Currently available 
options are "daily" (default) and "3-hourly".}

\item{dsn}{\code{character}, target folder for file download. Defaults to the 
current working directory if not specified otherwise.}

\item{format}{\code{character}, see \code{\link{strptime}}.}
}
\value{
A \code{data.frame} with downloaded .bin (or .Z for 3-hourly data) and .xml 
filepaths.
}
\description{
Download TRMM 3B42 daily or 3-hourly data for a given time span from the NASA 
FTP servers (\url{ftp://disc3.nascom.nasa.gov/data/s4pa/TRMM_L3}).
}
\examples{
 
\dontrun{
## download TRMM 3B42 daily data from Jan 1 to Jan 3, 2015
downloadTRMM(begin = "2015-01-01", end = "2015-01-03")

## same for 3-hourly data, from noon to noon
downloadTRMM(begin = "2015-01-01 12:00", end = "2015-01-03 12:00", 
             type = "3-hourly", format = "\%Y-\%m-\%d \%H:\%M")
}
              
}
\author{
Florian Detsch
}
\seealso{
\code{\link{strptime}}, \code{\link{download.file}}.
}

