% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadTRMM.R
\name{downloadTRMM}
\alias{downloadTRMM}
\title{Download TRMM 3B42 daily data}
\usage{
downloadTRMM(begin, end, dsn = ".", format = "\%Y-\%m-\%d")
}
\arguments{
\item{begin}{Date or character. Desired start date.}

\item{end}{Date or character. Desired end date.}

\item{dsn}{Character. Target directory for file download. If not supplied,
this defaults to the current working directory.}

\item{format}{Character. See \code{\link{as.Date}}.}

\item{...}{Further arguments. Currently not used.}
}
\value{
A vector of filepaths.
}
\description{
Download TRMM 3B42 daily binary data for a given time span from the NASA FTP servers
(\url{ftp://disc3.nascom.nasa.gov/data/s4pa/TRMM_L3/TRMM_3B42_daily/}).
}
\examples{
## download TRMM 3B42 data from Jan 1 to Jan 5, 2015
downloadTRMM(begin = "2015-01-01", end = "2015-01-05")
}
\author{
Florian Detsch
}
\seealso{
\code{\link{download.file}}, \code{\link{getURL}}
}

