% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendByDist.R
\name{extendByDist}
\alias{extendByDist}
\title{Extend an object's extent by distance}
\usage{
extendByDist(x, width)
}
\arguments{
\item{x}{Extent, Raster* or Spatial* object.}

\item{width}{Numeric. Distance to add to original geometry (in native map 
units). Can be of length 1 (applied equally in each direction), 2 (first 
entry applied in vertical direction, second entry in horizontal direction) or 
4 (applied separately in bottom/left/top/right direction).}
}
\value{
An extended extent.
}
\description{
Similar to \code{\link{extend}}, this function extends the extent of an 
Extent, Raster* or Spatial* object. Rather than requiring the user 
to supply either a target extent or a desired number of columns and rows to 
add to the initial Raster*, extension is based on native map units, e.g. 
decimal degrees when the projection is EPSG:4326.
}
\examples{
rst <- kiliAerial(rasterize = TRUE, minNumTiles = 4L)

extent(rst)
extendByDist(rst, width = 1000)

}
\author{
Florian Detsch
}
\seealso{
\code{\link{extend}}, \code{\link{extent}}
}

