% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentroids.R
\name{getCentroids}
\alias{getCentroids}
\title{Determine a polygon's centroid}
\usage{
getCentroids(spP)
}
\arguments{
\item{spP}{SpatialPolygons or SpatialPolygonsDataFrame.}
}
\value{
A numeric matrix of x and y coordinates.
}
\description{
Determine a polygon's centroid based on information stored in slot 'labpt'.
Taken from \url{http://linkedscience.org/tools/sparql-package-for-r/tutorial-part-ii-calculating-deforestation-per-state/}.
}
\seealso{
\code{\link{SpatialPolygons}}, \code{\link{SpatialPolygonsDataFrame}}
}
\author{
Florian Detsch
}
