% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/installEIPkgs.R
\name{installEIPkgs}
\alias{installEIPkgs}
\title{download and install packages for envin-umr}
\usage{
installEIPkgs(cran.pkgs = NULL, git.pkgs = NULL, check = TRUE,
  update = FALSE, MODIS = FALSE, dep = TRUE, ...)
}
\arguments{
\item{cran.pkgs}{character vector of packages to be installed from CRAN}

\item{git.pkgs}{character vector of packages to be installed from GitHub (not yet implemented)}

\item{check}{logical, whether to check if any packages specified in \code{cran.pkgs}
are already installed and skip their installation}

\item{update}{logical, whether to run \code{update.packages()} after installation}

\item{MODIS}{logical, if TRUE MODIS package will be installed from r-forge}

\item{dep}{logical, whether dependencies should be onstalled for specified packages}
}
\description{
download and install a bunch of packages that we find useful for our
day-to-day workflow. If \code{cran.pkgs = NULL} a standard list of
packages will be downloaded and installed. If you supply a character
vector of packages, these wil be installed instead.
}
\author{
Tim Appelhans
}

