% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiliAerial.R
\name{kiliAerial}
\alias{kiliAerial}
\title{Retrieve a Bing (or Google) Maps Aerial Image}
\usage{
kiliAerial(upperLeft, lowerRight, template = NULL,
  projection = "+init=epsg:4326", type = c("bing", "google"), ...)
}
\arguments{
\item{upperLeft, lowerRight}{\code{numeric}. Bounding box coordinates in the 
form \code{c(lat, long)}. If not specified and 'template' is missing, the 
resulting image defaults to the spatial extent of the study area covered in 
\strong{Detsch et al. (2016; see References)}.}

\item{template}{\code{Extent}, or any object from which an \code{Extent} can 
be extracted, see \code{\link[raster]{crop}}.}

\item{projection}{\code{character}, defaults to \code{"+init=epsg:4326"} 
(\url{http://spatialreference.org/ref/epsg/wgs-84/}). Desired coordinate 
reference system of the retrieved image.}

\item{type}{\code{character}. Tile server from which to download data. 
Currently available options are \code{"bing"} (default; only valid if 
\strong{OpenStreetMap} is available) and \code{google}.}

\item{...}{Additional arguments passed to \code{OpenStreetMap::openmap}.}
}
\value{
A 3-layered (i.e., RGB) \code{RasterStack} object.
}
\description{
Retrieve a Bing Maps aerial image centered on the Kilimanjaro region, 
Tanzania, by default. This image is aligned with the 1/12-degree AVHRR GIMMS 
NDVI grid covering the region. As an alternative, image retrieval is also 
available through Google Maps which is also switched on automatically if the 
\strong{OpenStreetMap} package is not available.
}
\examples{
 
\dontrun{
img <- kiliAerial(minNumTiles = 12L, projection = "+init=epsg:21037")
plotRGB(img)
}          
                   
}
\author{
Florian Detsch
}
\references{
Detsch F, Otte I, Appelhans T, Nauss T (2016) Seasonal and long-term 
vegetation dynamics from 1-km GIMMS-based NDVI time series at Mt. 
Kilimanjaro, Tanzania. Remote Sensing of Environment 178: 70-83. Available 
online at \url{http://dx.doi.org/10.1016/j.rse.2016.03.007}.
}
\seealso{
\code{openmap} and \code{openproj} in \strong{OpenStreetMap}, 
\code{\link{gmap}}.
}

