% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiliAerial.R
\name{kiliAerial}
\alias{kiliAerial}
\title{Retrieve an (aerial) image of Mt. Kilimanjaro}
\usage{
kiliAerial(upperLeft, lowerRight, template = NULL,
  projection = "+init=epsg:21037", type = c("bing", "google"), ...)
}
\arguments{
\item{upperLeft, lowerRight}{'numeric'. Upper-left (lower-right) bounding box 
coordinates passed on to \code{\link{openmap}}. If not supplied and 
'template' is missing, the returned image defaults to the extent of the study 
area covered in \strong{Detsch et al. (under review)}.}

\item{template}{'Extent' object, or any object that is compatible with 
\code{\link{extent}}.}

\item{projection}{'character', defaults to "+init=epsg:21037". Output projection
of the retrieved image, see \code{\link{openproj}}.}

\item{type}{'character', defaults to "bing". Tile server from which to get the
map.}

\item{...}{Further arguments passed on to \code{\link{openmap}}.}
}
\value{
An RGB 'RasterStack' object.
}
\description{
Based on functionalities offered by the R \bold{OpenStreetMap} package, 
this function per default retrieves a Bing aerial image centered on Mt. 
Kilimanjaro, Tanzania, that is aligned with the 8-km GIMMS NDVI grid covering 
the region.
}
\examples{
 
\dontrun{
## download high-resolution image
img <- kiliAerial(minNumTiles = 12L)

## create figure
plotRGB(img)
}          
                   
}
\author{
Florian Detsch
}
\references{
Detsch, F., Otte, I., Appelhans, T., and T. Nauss (2015): Seasonal and 
long-term vegetation dynamics from NDVI time series at Mt. Kilimanjaro, 
Tanzania. Under review at \emph{Remote Sensing of Environment}.
}
\seealso{
\code{\link{openmap}}, \code{\link{openproj}}, \code{\link{gmap}}.
}

