% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiliAerial.R
\name{kiliAerial}
\alias{kiliAerial}
\title{Retrieve an (aerial) image of Mt. Kilimanjaro}
\usage{
kiliAerial(upperLeft, lowerRight, template = NULL,
  projection = "+init=epsg:21037", type = "bing", rasterize = TRUE, ...)
}
\arguments{
\item{upperLeft, lowerRight}{'numeric'. Upper-left (lower-right) bounding box 
coordinates passed on to \code{\link{openmap}}. If not supplied and 
'template' is missing, the returned image defaults to the extent of the study 
area covered in \strong{Detsch et al. (under review)}.}

\item{template}{'Extent' object, or any object that is compatible with 
\code{\link{extent}}.}

\item{projection}{'character', defaults to "+init=epsg:21037". Output projection
of the retrieved image, see \code{\link{openproj}}.}

\item{type}{'character', defaults to "bing". Tile server from which to get the
map, see \code{\link{openmap}}.}

\item{rasterize}{'logical'. If \code{TRUE} (default), the function returns a 
RGB 'RasterStack' object.}

\item{...}{Further arguments passed on to \code{\link{openmap}}.}
}
\value{
An object of class \code{OpenStreetMap} or, if \code{rasterize = TRUE}, 
a multi-layer 'RasterStack' object.
}
\description{
Based on functionalities offered by the R \bold{OpenStreetMap} package, 
this function per default retrieves a Bing aerial image centered on Mt. 
Kilimanjaro, Tanzania, that is aligned with the 8-km GIMMS NDVI grid covering 
the region.
}
\examples{
 
## download high-resolution image
img <- kiliAerial(minNumTiles = 12L, rasterize = FALSE)

## create figure
library(ggplot2)
autoplot(img)
              
}
\author{
Florian Detsch
}
\references{
Detsch, F., Otte, I., Appelhans, T., and T. Nauss (2015): Seasonal and 
long-term vegetation dynamics from NDVI time series at Mt. Kilimanjaro, 
Tanzania. Under review at \emph{Remote Sensing of Environment}.
}
\seealso{
\code{\link{openmap}}, \code{\link{openproj}}
}

