% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapView.R
\docType{methods}
\name{mapView}
\alias{mapView}
\alias{mapView,RasterBrick-method}
\alias{mapView,RasterStack-method}
\alias{mapView,Satellite-method}
\alias{mapView,SpatialPointsDataFrame-method}
\alias{mapView,SpatialPolygonsDataFrame-method}
\title{view raster layers interactively}
\usage{
\S4method{mapView}{RasterLayer}(x, map = NULL, cols = envinmrPalette(7),
  na.color = "transparent", values = NULL, map.types = c("OpenStreetMap",
  "Esri.WorldImagery"), layer.opacity = 0.8, legend = TRUE,
  legend.opacity = 1, trim = TRUE, ...)

\S4method{mapView}{RasterStack}(x, ...)

\S4method{mapView}{RasterBrick}(x, ...)

\S4method{mapView}{SpatialPointsDataFrame}(x, map = NULL, burst = FALSE,
  cols = envinmrPalette(7), na.color = "transparent", values = NULL,
  map.types = c("OpenStreetMap", "Esri.WorldImagery"), layer.opacity = 0.8,
  legend = TRUE, legend.opacity = 1, ...)

\S4method{mapView}{SpatialPolygonsDataFrame}(x, map = NULL, burst = FALSE,
  cols = envinmrPalette(7), na.color = "transparent", values = NULL,
  map.types = c("OpenStreetMap", "Esri.WorldImagery"), layer.opacity = 0.8,
  legend = TRUE, legend.opacity = 1, weight = 2, ...)

\S4method{mapView}{Satellite}(x, ...)
}
\arguments{
\item{x}{a \code{\link{raster}}* object}

\item{cols}{color palette for the layers}

\item{na.color}{color for missing values}

\item{values}{a vector of values for the visualisation of the layers.
Per default these are calculated based on the supplied raster* object.}

\item{map.types}{character spcifications for the base maps.
see \url{http://leaflet-extras.github.io/leaflet-providers/preview/}
for available options.}

\item{layer.opacity}{opacity of the layers}

\item{legend}{should a legend be plotted}

\item{legend.opacity}{opacity of the legend}

\item{trim}{should the raster be trimmed in case there are NAs on the egdes}

\item{...}{additional arguments passed on to \code{\link{addLegend}}}

\item{burst}{whether to show all (TRUE) or only one (FALSE) layers}

\item{burst}{whether to show all (TRUE) or only one (FALSE) layers}
}
\description{
this function produces an interactive GIS-like view of the specified
raster layers on top of the specified base maps.
}
\section{Methods (by class)}{
\itemize{
\item \code{RasterStack}: 

\item \code{RasterBrick}: 

\item \code{SpatialPointsDataFrame}: 

\item \code{SpatialPolygonsDataFrame}: 

\item \code{Satellite}: 
}}
\examples{
### raster data ###
data(meuse.grid)
coordinates(meuse.grid) = ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
gridded(meuse.grid) = TRUE
meuse_rst <- stack(meuse.grid)

m1 <- mapView(meuse_rst)
m1

m2 <- mapView(meuse_rst[[1]])
m2

### point vector data ###
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")

# all layers of meuse
mapView(meuse)

# only one layer, all info in popups
mapView(meuse, burst = FALSE)

### overlay vector on top of raster ###
m3 <- mapView(meuse, map = m2, burst = FALSE)
m3

m4 <- mapView(meuse, map = m2)
m4

### polygon vector data ###
data("DEU_adm2")
mapView(gadm, burst = FALSE)
}
\author{
Tim Appelhans
}

