% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.filledcontour.R
\name{panel.filledcontour}
\alias{panel.filledcontour}
\title{create smooth contour fills for lattice levelplots}
\usage{
panel.filledcontour(x, y, z, subscripts, fill.contours = TRUE, at,
  contours = TRUE, at.contours = NULL, col.regions = cm.colors,
  col = col.regions(length(at) - 1), col.contours = "black",
  lty.contours = 1, lwd.contours = 1, draw.labels = TRUE, ...)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{z}{vector}

\item{subscripts}{optional}

\item{fill.contours}{should regions between contours be (colour) filled?}

\item{at}{sequence for the fill contours}

\item{contours}{should contour lines be drawn?}

\item{at.contours}{sequence for the contour lines}

\item{col.regions}{colour palette for filling}

\item{col}{set of colours passed on to \code{\link{filled.contour}}}

\item{col.contours}{colour of contour lines}

\item{lty.contours}{linetype of contour lines}

\item{lwd.contours}{size of contours}

\item{draw.labels}{should labels be drawn along the contour lines?}

\item{...}{currently not used}
}
\description{
this panel function will create a smooth fill for a 
\code{\link{levelplot}} similar to \code{\link{filled.contour}}
}
\examples{
library(latticeExtra)

data(volcano)

plot.new()
levelplot(volcano, panel = Rsenal:::panel.filledcontour,
          at.contour = seq(100, 200, 20))

}
\author{
Tim Appelhans 
(modified from: http://grokbase.com/t/r/r-help/089fhk4km2/r-creating-smooth-color-regions-with-panel-contourplot)
}

