% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/panel.map.R
\name{panel.map}
\alias{panel.map}
\title{panel function for maps}
\usage{
panel.map(map.data, fill = TRUE, col = "grey70", ...)
}
\arguments{
\item{map.data}{the map type - see \code{\link{map}} database argument}

\item{fill}{should the polygons be filled?}

\item{col}{the colour of the fill}

\item{...}{additional arguments passed on to \code{\link{panel.polygon}}}
}
\value{
A \code{\link{panel.polygon}} object
}
\description{
this function produces a \code{\link{panel.polygon}} from a
\code{\link{map}} object. This can then be used to visualise
the map on top of an \code{\link{spplot}} object.
NOTE: currently only tested for unprojected latlon data.
}
\examples{
library(latticeExtra)
rst <- raster(ncols = 360, nrows = 180)
rst[] <- rnorm(360 * 180)

spplot(rst) + layer(panel.map("world"))
}
\author{
Tim Appelhans
}

