% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotModelCV.R
\name{plotModelCV}
\alias{plotModelCV}
\title{plot cross-validated variable importance}
\usage{
plotModelCV(model, metric = model$metric, tuningValue = "Variables",
  xlim = "minmax", ylim = "minmax", sderror = FALSE, grid = TRUE)
}
\arguments{
\item{model}{A rfe or train object. See \code{\link{rfe}} and \code{\link{train}}}

\item{metric}{the metric to be used. Note this needs to be the metric used
to calculate the \code{\link{rfe}} or  \code{\link{train}} model}

\item{tuningValue}{The tuning value which is depicted on the x axis.
For rfe models default is "Variables", the number of variables.}

\item{xlim}{the xlim for the plot}

\item{ylim}{the ylim for the plot}

\item{sderror}{If TRUE then standard error is calculated. If FALSE then
standard deviations are used}

\item{grid}{Print grid or not}
}
\value{
a trellis object
}
\description{
this function plots variable imporatnce for cross-validated
\code{\link{rfe}} variable selection classes or \code{\link{train}} classes.
 It uses the k cross-validations to compute the mean +/- sd error or
 standard deviation metric.
}
\note{
if rfe is used as model, then returnResamp = "all"
must be set in rfe training
}
\author{
Hanna Meyer, Tim Appelhans
}

