% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwGen.R
\name{pwGen}
\alias{pwGen}
\title{generate passwords}
\usage{
pwGen(ndigits, seed = 123, master = NULL)
}
\arguments{
\item{ndigits}{desired number of digits}

\item{seed}{see \code{\link{set.seed}}}
}
\value{
a password.
}
\description{
generates a password based on a desired number of digits and a random seed.
}
\examples{
pwGen(4, seed = as.numeric(Sys.Date()))

}
\author{
Tim Appelhans
}
