% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_df.R
\name{quantile_df}
\alias{quantile_df}
\title{Pretty print quantiles}
\usage{
quantile_df(x, probs = seq(0, 1, 0.05), na.rm = F, names = F, type = 7,
  colname = NULL, round = NULL, ...)
}
\arguments{
\item{x}{numeric vector to calculate quantile with}

\item{probs}{numeric vector of probabilities to use for quantile}

\item{na.rm}{logical, pass to quantile function}

\item{names}{logical, pass to quantile function}

\item{type, }{number, pass to quantile function}

\item{colname, }{character, name of variable column in output}

\item{round, }{integer, number of digits to round quantile value.
passed to \code{digits} argument of \code{round} function}
}
\value{
prettified data.frame of probabilities
}
\description{
Useful for R Markdown reports
}
\examples{

quantile_df(mtcars$mpg, seq(0,1,.1))
quantile_df(mtcars$mpg, seq(0,1,.1), names=T)
quantile_df(mtcars$mpg, seq(0,1,.2), colname='mpg')
quantile_df(mtcars$mpg, round=1)
}

