% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeIMERG.R
\name{rasterizeIMERG}
\alias{rasterizeIMERG}
\title{Rasterize GPM IMERG data}
\usage{
rasterizeIMERG(x, layer = "precipitationCal")
}
\arguments{
\item{x}{Filename of the IMERG HDF5 file}

\item{layer}{IMERG layer name}
}
\value{
A \code{RasterLayer}
}
\description{
Data can be downloaded in HDF5 from 
http://pmm.nasa.gov/data-access/downloads/gpm
This function mainly follows the description for ArcGIS provided by the NASA: 
http://disc.sci.gsfc.nasa.gov/recipes/?q=recipes/How-to-Import-HDF5-formatted-IMERG-GPM-Precipitation-Data-into-ArcGIS
}
\author{
Hanna Meyer
}

