% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rasterizeTRMM.R
\name{rasterizeTRMM}
\alias{rasterizeTRMM}
\title{Rasterize TRMM 3B42 binary data}
\usage{
rasterizeTRMM(binary, meta, write_out = TRUE)
}
\arguments{
\item{binary}{Character. TRMM binary file (*.bin).}

\item{meta}{Character. Corresponding metadata file (*.xml).}

\item{write_out}{Logical. Enables output storage.}
}
\value{
A \code{RasterLayer} object.
}
\description{
Rasterize TRMM 3B42 daily binary data, e.g. previously downloaded via
\code{\link{downloadTRMM}}.
}
\examples{
## download TRMM 3B42 data from Jan 1 to Jan 5, 2015
fls_trmm <- downloadTRMM(begin = "2015-01-01", end = "2015-01-05")

## rasterize
rst_trmm <- rasterizeTRMM(binary = fls_trmm[1, 1], meta = fls_trmm[1, 2],
                          write_out = FALSE)

plot(rst_trmm)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{readBin}}, \code{\link{downloadTRMM}}
}

