% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNA.R
\name{removeNA}
\alias{removeNA}
\title{remove rows with NA values in selcted columns of a data frame}
\usage{
removeNA(df, cols)
}
\arguments{
\item{df}{a data frame}

\item{cols}{a vector of column names or column numbers}
}
\value{
A data frame
}
\description{
this function will remove all NA entries from selected columns of a data frame
}
\examples{
df <- data.frame(a = 1:10, b = letters[1:10], c = rnorm(10, 2, 1))
df$a[sample(nrow(df), 3)] <- NA
df$b[sample(nrow(df), 2)] <- NA
df$c[sample(nrow(df), 1)] <- NA

df

removeNA(df, 1)
removeNA(df, 2)
removeNA(df, 3)
removeNA(df, c(1, 2))
removeNA(df, c(1, 2, 3))

}
\author{
Tim Appelhans
}
