% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgb2spLayout.R
\name{rgb2spLayout}
\alias{rgb2spLayout}
\title{Convert an RGB RasterBrick/Stack to use with spplot}
\usage{
rgb2spLayout(x, alpha)
}
\arguments{
\item{x}{a RasterBrick or RasterStack}
}
\description{
this function takes a RGB image stack/brick and produces a list with
the colorinformation which can be used in the sp.layout argument of spplot
}
\examples{
library(raster)
library(sp)

b <- brick(system.file("external/rlogo.grd", package="raster"))

## using plotRGB
plotRGB(b)

## convert brick to list
lout <- rgb2spLayout(b)
lout_alph <- rgb2spLayout(b, alpha = 0.5)

## create random spatial points for plotting
df <- data.frame(dat = rnorm(100, 2, 1),
                 x = rnorm(100, 50, 20),
                 y = rnorm(100, 50, 25))
coordinates(df) <- ~x+y

## plot spatial points with rgb background
spplot(df, sp.layout = lout)
spplot(df, sp.layout = lout_alph)
}
\author{
Tim Appelhans
}
\seealso{
\code{\link{plotRGB}}
}

