% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rsenal-package.R
\docType{package}
\name{Rsenal-package}
\alias{Rsenal}
\alias{Rsenal-package}
\title{Rsenal: magic functions for things various}
\description{
an Rsenal of functions by Environmentalinformatics Marburg
}
\details{
A collection of functions which we find helpful/useful for our
day-to-day ventures into data analysis at Environmentalinformatics
group, Department of Geography, Philipps University Marburg.
There is no specific focus of interest in this collection, some
functions are very task specific (e.g. \code{\link{roc}} which let's
you create roc curves for land-use classifications), yet others are
very general and widely applicable (e.g. \code{\link{latticeCombineGrid}}
which produces a panel plot based on a list of lattice plot objects).
}
\author{
Tim Appelhans, Florian Detsch, Insa Otte, Meike Kuehnlein,
Hanna Meyer, Spaska Forteva, Thomas Nauss, Stefan Woellauer\cr
\cr
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\keyword{package}

