% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shinyTable}
\alias{shinyTable}
\title{Shiny app to visualize dataframe as simple interactive datatable}
\usage{
shinyTable(df)
}
\arguments{
\item{df}{dataframe to be visualized}
}
\value{
Shiny App
}
\description{
Launches a basic Shiny App that renders the given dataframe into an interactive datatable using \code{renderDataTable}
}
\examples{
\dontrun{
shinyTable(mtcars)
}
}

