% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitRaster.R
\name{splitRaster}
\alias{splitRaster}
\title{Split a Raster object}
\usage{
splitRaster(file, s = 2, ...)
}
\arguments{
\item{file}{Character. Filename.}

\item{s}{Integer. Number of slices \code{file} should be divided into.}

\item{...}{Further arguments passed on to \code{\link{gdal_translate}}.}
}
\value{
A list of RasterBrick objects.
}
\description{
Split a Raster* object into a given number of horizontal and vertical slices.
Inspired by \url{http://java.dzone.com/articles/use-gdal-r-console-split}.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{gdal_translate}}, \code{\link{GDALinfo}}
}

