% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvi.R
\name{vvi}
\alias{vvi}
\alias{vvi,missing-method}
\alias{vvi,Raster-method}
\alias{vvi,stars-method}
\title{Visible vegetation index}
\usage{
\S4method{vvi}{missing}(r = 1, g = 2, b = 3)

\S4method{vvi}{Raster}(x, r = 1, g = 2, b = 3)

\S4method{vvi}{stars}(x, r = 1, g = 2, b = 3)
}
\arguments{
\item{r, g, b}{Band numbers as \code{integer}.}

\item{x}{A \code{Raster} or \code{stars} object.}
}
\value{
A VVI \code{RasterLayer} or \code{stars} object.
}
\description{
This function calculates a visible vegetation index (VVI) from a RGB 
\code{Raster*} object.
}
\examples{
library(RColorBrewer)

data(gmap_hel)
plotRGB(gmap_hel)

gmap_hel_veg = vvi(gmap_hel)
plot(gmap_hel_veg, col = brewer.pal(5, "BrBG"), alpha = .5, add = TRUE)

}
\references{
Planetary Habitability Laboratory (2015): Visible Vegetation Index (VVI). 
Available online: \url{http://phl.upr.edu/projects/visible-vegetation-index-vvi}.
}
\author{
Florian Detsch, Tim Appelhans
}
